% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{add_ode_compartment}
\alias{add_ode_compartment}
\title{Add ODE compartment to compartments object.}
\usage{
add_ode_compartment(compartments, ode)
}
\arguments{
\item{compartments}{compartments object}

\item{ode}{ODE}
}
\value{
a compartments object
}
\description{
Add ODE compartment to compartments object.
}
\keyword{internal}
