% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{addEquation}
\alias{addEquation}
\title{Add equation.}
\usage{
addEquation(object, lhs, rhs, ...)
}
\arguments{
\item{object}{object containing code record(s)}

\item{lhs}{left-hand-side variable}

\item{rhs}{right-hand-side expression}

\item{...}{extra arguments like before or after}
}
\value{
updated object
}
\description{
Add equation.
}
