% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{addInitialCondition}
\alias{addInitialCondition}
\title{Add initial condition to compartments object.}
\usage{
addInitialCondition(line, emptyCondition, compartments)
}
\arguments{
\item{line}{line record}

\item{emptyCondition}{empty condition, to be completed}

\item{compartments}{compartments object}
}
\value{
updated compartments object
}
\description{
Add initial condition to compartments object.
}
