% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{addODECompartment}
\alias{addODECompartment}
\title{Add ODE compartment to comparments object.}
\usage{
addODECompartment(compartments, ode)
}
\arguments{
\item{compartments}{compartments object}

\item{ode}{ODE}
}
\value{
a compartments object
}
\description{
Add ODE compartment to comparments object.
}
