% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{addParameter}
\alias{addParameter}
\title{Add a parameter in list.}
\usage{
addParameter(object, parameter)
}
\arguments{
\item{object}{generic object}

\item{parameter}{parameter to add}
}
\value{
filtered object
}
\description{
Add a parameter in list.
}
