% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{addProperties}
\alias{addProperties}
\title{Add properties to compartments object.}
\usage{
addProperties(compartments, records, name, init)
}
\arguments{
\item{compartments}{compartments object}

\item{records}{all records}

\item{name}{record name to look at}

\item{init}{empty characteristic, to be completed}
}
\value{
updated compartments object
}
\description{
Add properties to compartments object.
}
\keyword{internal}
