% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/parameters.R, R/campsis_model.R
\name{addRSE}
\alias{addRSE}
\alias{addRSE,parameters,parameter,numeric-method}
\alias{addRSE,campsis_model,parameter,numeric-method}
\title{Add relative standard error (RSE) to the specified parameter.}
\usage{
addRSE(object, parameter, value, ...)

\S4method{addRSE}{parameters,parameter,numeric}(object, parameter, value, ...)

\S4method{addRSE}{campsis_model,parameter,numeric}(object, parameter, value, ...)
}
\arguments{
\item{object}{model or parameters object}

\item{parameter}{parameter object (Theta, Omega or Sigma)}

\item{value}{RSE value, in percent}

\item{...}{extra arguments, unused}
}
\value{
updated object
}
\description{
Add relative standard error (RSE) to the specified parameter.
}
