% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{addSuffix}
\alias{addSuffix}
\title{Generic function to add a suffix to various objects like parameters, code records,
compartment names or a model (all previous objects at the same time).
This makes it an extremely powerful function to combine 2 models or more (using function 'add'),
that have similar equation, parameter or compartment names.}
\usage{
addSuffix(object, suffix, separator, ...)
}
\arguments{
\item{object}{generic object}

\item{suffix}{suffix to be appended, single character value}

\item{separator}{separator to use before the suffix, default is the underscore}

\item{...}{extra arguments like 'model' if the changes need to be reflected in the model}
}
\value{
updated object of the same class as the provided object, unless 'model' was specified, in that case the model is returned
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`addSuffix()` is deprecated in favor of `add_suffix()`.
}
\keyword{internal}
