% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{all_na}
\alias{all_na}
\title{Check is vector has NA's only.}
\usage{
all_na(x)
}
\arguments{
\item{x}{any vector}
}
\value{
TRUE if all values are NA, FALSE otherwise
}
\description{
Check is vector has NA's only.
}
\keyword{internal}
