% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{append_code_records}
\alias{append_code_records}
\title{Append code records}
\usage{
append_code_records(records1, records2)
}
\arguments{
\item{records1}{base set of code records}

\item{records2}{extra set of code records to be appended}
}
\value{
the resulting set of code records
}
\description{
Append code records
}
\keyword{internal}
