% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_statement.R
\name{append_comment}
\alias{append_comment}
\title{Append comment.}
\usage{
append_comment(str, object, dest)
}
\arguments{
\item{str}{single character string}

\item{object}{model statement}

\item{dest}{destination engine, string}
}
\value{
single character string followed by comment
}
\description{
Append comment.
}
\keyword{internal}
