% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{appendComma}
\alias{appendComma}
\title{Append comma at the right place. The goal of this function is to preserve
comments in mrgsolve as well. Comma has to be placed before the comment '#'.}
\usage{
appendComma(x)
}
\arguments{
\item{x}{record line}
}
\value{
same line with a comma at the right place
}
\description{
Append comma at the right place. The goal of this function is to preserve
comments in mrgsolve as well. Comma has to be placed before the comment '#'.
}
