% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_statement.R
\name{appendComment}
\alias{appendComment}
\title{Append comment.}
\usage{
appendComment(str, object, dest)
}
\arguments{
\item{str}{single character string}

\item{object}{model statement}

\item{dest}{destination engine, string}
}
\description{
Append comment.
}
