% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{appendCompartments}
\alias{appendCompartments}
\title{Append compartments.}
\usage{
appendCompartments(compartments1, compartments2)
}
\arguments{
\item{compartments1}{base set of compartments}

\item{compartments2}{extra set of compartments to be appended}
}
\value{
the resulting set of compartments
}
\description{
Append compartments.
}
\keyword{internal}
