% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_model.R
\name{appendModel}
\alias{appendModel}
\title{Append model (or simply add).}
\usage{
appendModel(model1, model2)
}
\arguments{
\item{model1}{base model}

\item{model2}{model to append}
}
\value{
the resulting CAMPSIS model
}
\description{
Append model (or simply add).
}
