% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{appendParameters}
\alias{appendParameters}
\title{Append parameters.}
\usage{
appendParameters(params1, params2)
}
\arguments{
\item{params1}{base set of parameters}

\item{params2}{extra set of parameters to be appended}
}
\value{
the resulting set of parameters
}
\description{
Append parameters.
}
