% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{as.data.frame}
\alias{as.data.frame}
\title{As data frame method.}
\usage{
as.data.frame(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{generic object}

\item{row.names}{row names}

\item{optional}{optional}

\item{...}{extra arguments}
}
\value{
data frame
}
\description{
As data frame method.
}
