% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assert_single_character_string}
\alias{assert_single_character_string}
\title{Assert the given character vector is a single character string.}
\usage{
assert_single_character_string(x)
}
\arguments{
\item{x}{single character string}
}
\value{
no return value
}
\description{
Assert the given character vector is a single character string.
}
