% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assertSingleCharacterString}
\alias{assertSingleCharacterString}
\title{Assert the given character vector is a single character string.}
\usage{
assertSingleCharacterString(x)
}
\arguments{
\item{x}{single character string}
}
\value{
no return value
}
\description{
Assert the given character vector is a single character string.
}
