% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/campsis_model.R
\name{auto_detect_nonmem}
\alias{auto_detect_nonmem}
\alias{auto_detect_nonmem,campsis_model-method}
\title{Auto-detect special variables from NONMEM as compartment properties.
Bioavailabilities, infusion durations/rates and lag times will be automatically
detected.}
\usage{
auto_detect_nonmem(object, ...)

\S4method{auto_detect_nonmem}{campsis_model}(object, ...)
}
\arguments{
\item{object}{object that has NONMEM special variables to be identified}

\item{...}{extra arguments, unused}
}
\value{
updated object
}
\description{
Auto-detect special variables from NONMEM as compartment properties.
Bioavailabilities, infusion durations/rates and lag times will be automatically
detected.
}
