% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/campsis_model.R
\name{autoDetectNONMEM}
\alias{autoDetectNONMEM}
\alias{autoDetectNONMEM,campsis_model-method}
\title{Auto-detect special variables from NONMEM to fill in compartment properties.
Bioavailabilities, infusion durations/rates and lag times will be automatically
detected.}
\usage{
autoDetectNONMEM(object, ...)

\S4method{autoDetectNONMEM}{campsis_model}(object, ...)
}
\arguments{
\item{object}{object that has NONMEM special variables to be identified}
}
\value{
updated object
}
\description{
Auto-detect special variables from NONMEM to fill in compartment properties.
Bioavailabilities, infusion durations/rates and lag times will be automatically
detected.
}
