% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campsis_model.R
\docType{class}
\name{campsis_model-class}
\alias{campsis_model-class}
\title{Campsis model class.}
\description{
Campsis model class.
}
\section{Slots}{

\describe{
\item{\code{model}}{a list of code records}

\item{\code{parameters}}{model parameters}

\item{\code{compartments}}{model compartments}
}}

