% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campsis_model.R
\name{CampsisModel}
\alias{CampsisModel}
\title{Create a new Campsis model.}
\usage{
CampsisModel(json = NULL)
}
\arguments{
\item{json}{path to JSON Campsis model file or JSON content in string form}
}
\value{
a Campsis model, empty
}
\description{
Create a new Campsis model.
}
