% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{checkMatrixIsPositiveDefinite}
\alias{checkMatrixIsPositiveDefinite}
\title{Check OMEGA/SIGMA matrix for positive definiteness.}
\usage{
checkMatrixIsPositiveDefinite(table, parameters)
}
\arguments{
\item{table}{data frame with the sampled parameters to check}

\item{parameters}{double array parameters to check for positive definiteness}
}
\description{
Check OMEGA/SIGMA matrix for positive definiteness.
}
\keyword{internal}
