% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{clean}
\alias{clean}
\alias{clean,parameters-method}
\title{Clean}
\usage{
clean(object)

\S4method{clean}{parameters}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
cleaned object
}
\description{
Clean
}
