% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_record.R
\docType{class}
\name{code_record-class}
\alias{code_record-class}
\title{Code record class. See this code record as an abstract class.
2 implementations are possible:
- properties record (lag, duration, rate & bioavailability properties)
- statements record (main, ode & error records)}
\description{
Code record class. See this code record as an abstract class.
2 implementations are possible:
- properties record (lag, duration, rate & bioavailability properties)
- statements record (main, ode & error records)
}
\section{Slots}{

\describe{
\item{\code{comment}}{a comment, single character value}

\item{\code{statements}}{model statements}
}}

