% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comment.R
\docType{class}
\name{comment-class}
\alias{comment-class}
\title{Comment class. A statement starting with #.}
\description{
Comment class. A statement starting with #.
}
\section{Slots}{

\describe{
\item{\code{comment}}{a comment associated to this model statement}
}}

