% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.R
\docType{class}
\name{compartment-class}
\alias{compartment-class}
\title{Compartment class.}
\description{
Compartment class.
}
\section{Slots}{

\describe{
\item{\code{name}}{compartment name (without prefix)}

\item{\code{index}}{compartment index}
}}

