% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.R
\name{Compartment}
\alias{Compartment}
\title{Create a compartment.}
\usage{
Compartment(index, name = NA)
}
\arguments{
\item{index}{compartment index}

\item{name}{compartment name}
}
\value{
an empty list of compartments
}
\description{
Create a compartment.
}
