% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_bioavailability.R
\docType{class}
\name{compartment_bioavailability-class}
\alias{compartment_bioavailability-class}
\title{Compartment bioavailability class.}
\description{
Compartment bioavailability class.
}
