% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_infusion_duration.R
\docType{class}
\name{compartment_infusion_duration-class}
\alias{compartment_infusion_duration-class}
\title{Compartment infusion duration class.}
\description{
Compartment infusion duration class.
}
