% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_initial_condition.R
\docType{class}
\name{compartment_initial_condition-class}
\alias{compartment_initial_condition-class}
\title{Compartment initial condition class.}
\description{
Compartment initial condition class.
}
