% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_property.R
\docType{class}
\name{compartment_property-class}
\alias{compartment_property-class}
\title{Compartment property class.}
\description{
Compartment property class.
}
\section{Slots}{

\describe{
\item{\code{compartment}}{related compartment index}

\item{\code{rhs}}{right-hand side formula}

\item{\code{comment}}{comment if any, single character string}
}}

