% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_lag_time.R
\name{CompartmentLagTime}
\alias{CompartmentLagTime}
\title{Create a lag time for the specified compartment.}
\usage{
CompartmentLagTime(compartment, rhs)
}
\arguments{
\item{compartment}{compartment index}

\item{rhs}{right-hand side part of the equation}
}
\value{
lag time
}
\description{
Create a lag time for the specified compartment.
}
