% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\docType{class}
\name{compartments-class}
\alias{compartments-class}
\title{Compartments class.}
\description{
Compartments class.
}
\section{Slots}{

\describe{
\item{\code{properties}}{compartment properties of the compartments defined in this class}
}}

