% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.R, R/compartments.R
\name{Compartments}
\alias{Compartments}
\title{Create a compartment.}
\usage{
Compartments()

Compartments()
}
\value{
an empty list of compartments

an empty list of compartments
}
\description{
Create a compartment.

Create a list of compartments
}
