% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R, R/compartments.R,
%   R/code_record.R
\name{contains}
\alias{contains}
\alias{contains,pmx_list,pmx_element-method}
\alias{contains,compartments,compartment_property-method}
\alias{contains,statements_record,model_statement-method}
\title{Check if an element exists in list.}
\usage{
contains(object, x)

\S4method{contains}{pmx_list,pmx_element}(object, x)

\S4method{contains}{compartments,compartment_property}(object, x)

\S4method{contains}{statements_record,model_statement}(object, x)
}
\arguments{
\item{object}{list object}

\item{x}{element to check if exists}
}
\value{
logical value, TRUE or FALSE
}
\description{
Check if an element exists in list.
}
