% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{convert_any_comment}
\alias{convert_any_comment}
\title{Convert Campsis comment style to C/C++ code.
Only the first # is translated to //.}
\usage{
convert_any_comment(x)
}
\arguments{
\item{x}{any record line}
}
\value{
same line with comments translated to C/C++
}
\description{
Convert Campsis comment style to C/C++ code.
Only the first # is translated to //.
}
\keyword{internal}
