% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{convert_outvars_to_capture}
\alias{convert_outvars_to_capture}
\title{Convert outvars argument to capture. Variables that are already in error block
will be discarded.}
\usage{
convert_outvars_to_capture(outvars, model)
}
\arguments{
\item{outvars}{character vector}

\item{model}{Campsis model}
}
\value{
all variables to capture
}
\description{
Convert outvars argument to capture. Variables that are already in error block
will be discarded.
}
\keyword{internal}
