% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{convertAnyComment}
\alias{convertAnyComment}
\title{Convert CAMPSIS comment style to C/C++ code.
Only the first # is translated to //.}
\usage{
convertAnyComment(x)
}
\arguments{
\item{x}{any record line}
}
\value{
same line with comments translated to C/C++
}
\description{
Convert CAMPSIS comment style to C/C++ code.
Only the first # is translated to //.
}
