% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{convertOutvarsToCapture}
\alias{convertOutvarsToCapture}
\title{Convert outvars argument to capture. Variables that are already in error block
will be discarded.}
\usage{
convertOutvarsToCapture(outvars, model)
}
\arguments{
\item{outvars}{character vector}

\item{model}{PMX model}
}
\value{
all variables to capture
}
\description{
Convert outvars argument to capture. Variables that are already in error block
will be discarded.
}
