% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R, R/compartments.R,
%   R/parameters.R, R/code_record.R, R/code_records.R, R/campsis_model.R
\name{delete}
\alias{delete}
\alias{delete,pmx_list,pmx_element-method}
\alias{delete,pmx_list,integer-method}
\alias{delete,compartments,compartment_property-method}
\alias{delete,parameters,single_array_parameter-method}
\alias{delete,parameters,double_array_parameter-method}
\alias{delete,statements_record,model_statement-method}
\alias{delete,code_records,model_statement-method}
\alias{delete,campsis_model,compartment_property-method}
\alias{delete,campsis_model,parameter-method}
\alias{delete,campsis_model,code_record-method}
\alias{delete,campsis_model,model_statement-method}
\title{Delete an element from this list.}
\usage{
delete(object, x)

\S4method{delete}{pmx_list,pmx_element}(object, x)

\S4method{delete}{pmx_list,integer}(object, x)

\S4method{delete}{compartments,compartment_property}(object, x)

\S4method{delete}{parameters,single_array_parameter}(object, x)

\S4method{delete}{parameters,double_array_parameter}(object, x)

\S4method{delete}{statements_record,model_statement}(object, x)

\S4method{delete}{code_records,model_statement}(object, x)

\S4method{delete}{campsis_model,compartment_property}(object, x)

\S4method{delete}{campsis_model,parameter}(object, x)

\S4method{delete}{campsis_model,code_record}(object, x)

\S4method{delete}{campsis_model,model_statement}(object, x)
}
\arguments{
\item{object}{list object}

\item{x}{element to delete or element index}
}
\value{
the updated list
}
\description{
Delete an element from this list.
}
