% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{disable}
\alias{disable}
\title{Disable.}
\usage{
disable(object, x, ...)
}
\arguments{
\item{object}{generic object}

\item{x}{what needs to be disabled}

\item{...}{extra arguments needed for disabling}
}
\value{
standardised object
}
\description{
Disable.
}
