% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\name{discard}
\alias{discard}
\alias{discard,ANY-method}
\title{Discard element(s) from list using predicate function.}
\usage{
discard(.x, .p, ...)

\S4method{discard}{ANY}(.x, .p, ...)
}
\arguments{
\item{.x}{a list, a vector or a PMX list}

\item{.p}{a predicate function}

\item{...}{extra arguments, ignored}
}
\value{
updated list, vector or PMX list
}
\description{
Discard element(s) from list using predicate function.
}
