% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_equation.R
\docType{class}
\name{equation-class}
\alias{equation-class}
\title{Equation class. Any statement in the form A = B.}
\description{
Equation class. Any statement in the form A = B.
}
\section{Slots}{

\describe{
\item{\code{lhs}}{left-hand side expression}

\item{\code{rhs}}{right-hand side expression}
}}

