% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_equation.R
\name{Equation}
\alias{Equation}
\title{Create a new equation.}
\usage{
Equation(lhs, rhs = "", comment = as.character(NA))
}
\arguments{
\item{lhs}{left-hand side variable corresponding to the assigned variable name}

\item{rhs}{right-hand side expression corresponding to a formula}

\item{comment}{comment if any, single character string}
}
\value{
an equation
}
\description{
Create a new equation.
}
