% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/pmx_model.R, R/rxode_model.R,
%   R/mrgsolve_model.R
\name{export}
\alias{export}
\alias{export,pmx_model,character-method}
\alias{export,pmx_model,rxode_type-method}
\alias{export,pmx_model,mrgsolve_type-method}
\title{Export function.}
\usage{
export(object, dest, ...)

\S4method{export}{pmx_model,character}(object, dest, outvars = NULL)

\S4method{export}{pmx_model,rxode_type}(object, dest, ...)

\S4method{export}{pmx_model,mrgsolve_type}(object, dest, outvars = NULL)
}
\arguments{
\item{object}{generic object}

\item{dest}{destination}

\item{...}{optional arguments}

\item{outvars}{additional variables to capture}
}
\value{
specific object depending on given destination
}
\description{
Export function.
}
