% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/parameter.R, R/parameters.R,
%   R/campsis_model.R
\name{export_to_json}
\alias{export_to_json}
\alias{export_to_json,theta-method}
\alias{export_to_json,omega-method}
\alias{export_to_json,sigma-method}
\alias{export_to_json,parameters-method}
\alias{export_to_json,campsis_model-method}
\title{Export the given object to a JSON object, ready to be written to files.}
\usage{
export_to_json(object, ...)

\S4method{export_to_json}{theta}(object, ...)

\S4method{export_to_json}{omega}(object, ...)

\S4method{export_to_json}{sigma}(object, ...)

\S4method{export_to_json}{parameters}(object, ...)

\S4method{export_to_json}{campsis_model}(object, ...)
}
\arguments{
\item{object}{any object}

\item{...}{extra arguments, unused}
}
\value{
the loaded S4 object
}
\description{
Export the given object to a JSON object, ready to be written to files.
}
