% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/parameter.R, R/parameters.R,
%   R/campsis_model.R
\name{exportToJSON}
\alias{exportToJSON}
\alias{exportToJSON,theta-method}
\alias{exportToJSON,omega-method}
\alias{exportToJSON,sigma-method}
\alias{exportToJSON,parameters-method}
\alias{exportToJSON,campsis_model-method}
\title{Export the given object to a JSON object, ready to be written to files.}
\usage{
exportToJSON(object, ...)

\S4method{exportToJSON}{theta}(object, ...)

\S4method{exportToJSON}{omega}(object, ...)

\S4method{exportToJSON}{sigma}(object, ...)

\S4method{exportToJSON}{parameters}(object, ...)

\S4method{exportToJSON}{campsis_model}(object, ...)
}
\arguments{
\item{object}{any object}

\item{...}{extra arguments, unused}
}
\value{
the loaded S4 object
}
\description{
Export the given object to a JSON object, ready to be written to files.
}
