% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_lhs}
\alias{extract_lhs}
\title{Extract left-hand-side expression.}
\usage{
extract_lhs(x, split = "=")
}
\arguments{
\item{x}{character value}

\item{split}{character where to split}
}
\value{
left-hand-side expression, not trimmed
}
\description{
Extract left-hand-side expression.
}
