% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_rhs}
\alias{extract_rhs}
\title{Extract right-hand-side expression.}
\usage{
extract_rhs(x, split = "=")
}
\arguments{
\item{x}{character value}

\item{split}{character where to split}
}
\value{
right-hand side expression
}
\description{
Extract right-hand-side expression.
}
