% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_text_between_brackets}
\alias{extract_text_between_brackets}
\title{Extract text between brackets.}
\usage{
extract_text_between_brackets(x)
}
\arguments{
\item{x}{character value}
}
\value{
text between brackets (trimmed)
}
\description{
Extract text between brackets.
}
