% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{extractCharacteristicsFromDesRecord}
\alias{extractCharacteristicsFromDesRecord}
\title{Extract all compartment characteristics from the DES record.}
\usage{
extractCharacteristicsFromDesRecord(record)
}
\arguments{
\item{record}{DES record}
}
\value{
a list
}
\description{
Extract all compartment characteristics from the DES record.
}
