% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{extractModelParametersFromNames}
\alias{extractModelParametersFromNames}
\title{Extract model parameters based on parameter names.}
\usage{
extractModelParametersFromNames(parameters, names)
}
\arguments{
\item{parameters}{all Campsis parameters}

\item{names}{names of the parameters to extract}
}
\value{
subset of Campsis parameters
}
\description{
Extract model parameters based on parameter names.
}
