% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extractRhs}
\alias{extractRhs}
\title{Extract right-hand-side expression.}
\usage{
extractRhs(x, split = "=")
}
\arguments{
\item{x}{character value}

\item{split}{character where to split}
}
\value{
right-hand side expressionn
}
\description{
Extract right-hand-side expression.
}
