% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extractTextBetweenBrackets}
\alias{extractTextBetweenBrackets}
\title{Extract text between brackets.}
\usage{
extractTextBetweenBrackets(x)
}
\arguments{
\item{x}{character value}
}
\value{
text between brackets (trimmed)
}
\description{
Extract text between brackets.
}
