% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{fix_omega}
\alias{fix_omega}
\alias{fix_omega,parameters-method}
\title{Fix omega matrix for SAME OMEGA parameters that have NA values due to imperfections in Pharmpy import.}
\usage{
fix_omega(object)

\S4method{fix_omega}{parameters}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
the parameter that matches
}
\description{
Fix omega matrix for SAME OMEGA parameters that have NA values due to imperfections in Pharmpy import.
}
