% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{flagOutOfRangeParameterRows}
\alias{flagOutOfRangeParameterRows}
\title{Flag all parameter rows that have at least one parameter out of the specified range.}
\usage{
flagOutOfRangeParameterRows(table, minMaxMean)
}
\arguments{
\item{table}{a data frame returned by sampleMore}

\item{minMaxMean}{a data frame with min, max and mean values for each parameter}
}
\description{
Flag all parameter rows that have at least one parameter out of the specified range.
}
