% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{flagSampledParameterRows}
\alias{flagSampledParameterRows}
\title{Flag all parameter rows that have at least one parameter out of the specified range
or that have a non positive definite OMEGA or SIGMA matrix.}
\usage{
flagSampledParameterRows(table, minMax, settings, parameters)
}
\arguments{
\item{table}{a data frame returned by \code{sampleMore}}

\item{minMax}{a data frame with min, max values for each parameter}

\item{settings}{replication settings}

\item{parameters}{double array parameters to check for positive definiteness}
}
\description{
Flag all parameter rows that have at least one parameter out of the specified range
or that have a non positive definite OMEGA or SIGMA matrix.
}
\keyword{internal}
