% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R, R/compartment_properties.R,
%   R/parameters.R
\name{get_by_index}
\alias{get_by_index}
\alias{get_by_index,pmx_list,integer-method}
\alias{get_by_index,pmx_list,numeric-method}
\alias{get_by_index,compartment_properties,compartment_property-method}
\alias{get_by_index,parameters,parameter-method}
\title{Get element by index.}
\usage{
get_by_index(object, x)

\S4method{get_by_index}{pmx_list,integer}(object, x)

\S4method{get_by_index}{pmx_list,numeric}(object, x)

\S4method{get_by_index}{compartment_properties,compartment_property}(object, x)

\S4method{get_by_index}{parameters,parameter}(object, x)
}
\arguments{
\item{object}{list object}

\item{x}{element index}
}
\value{
element from the list whose index matches with provided index
}
\description{
Get element by index.
}
