% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\name{get_by_name}
\alias{get_by_name}
\alias{get_by_name,pmx_list,character-method}
\title{Get an element from a list by name.
Never return more than 1 element.}
\usage{
get_by_name(object, name)

\S4method{get_by_name}{pmx_list,character}(object, name)
}
\arguments{
\item{object}{list object}

\item{name}{element name to search for}
}
\value{
the element that was found or NULL if no element was found with the same name
}
\description{
Get an element from a list by name.
Never return more than 1 element.
}
