% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_campsismod_options}
\alias{get_campsismod_options}
\title{Get the Campsismod options (R options).}
\usage{
get_campsismod_options()
}
\value{
global options for Campsismod
}
\description{
Get the Campsismod options (R options).
}
\keyword{internal}
