% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/compartments.R, R/campsis_model.R
\name{get_compartment_index}
\alias{get_compartment_index}
\alias{get_compartment_index,compartments,character-method}
\alias{get_compartment_index,campsis_model,character-method}
\title{Get the compartment index for the specified compartment name.}
\usage{
get_compartment_index(object, name)

\S4method{get_compartment_index}{compartments,character}(object, name)

\S4method{get_compartment_index}{campsis_model,character}(object, name)
}
\arguments{
\item{object}{generic object that contains compartments information}

\item{name}{compartment name}
}
\value{
the corresponding compartment index
}
\description{
Get the compartment index for the specified compartment name.
}
