% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{get_name_in_model}
\alias{get_name_in_model}
\alias{get_name_in_model,theta-method}
\alias{get_name_in_model,omega-method}
\alias{get_name_in_model,sigma-method}
\title{Get the name of the given parameter in the Campsis model.}
\usage{
get_name_in_model(x)

\S4method{get_name_in_model}{theta}(x)

\S4method{get_name_in_model}{omega}(x)

\S4method{get_name_in_model}{sigma}(x)
}
\arguments{
\item{x}{element to know the name}
}
\value{
the name of this parameter
}
\description{
Get the name of the given parameter in the Campsis model.
}
