% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\name{get_names}
\alias{get_names}
\alias{get_names,pmx_list-method}
\title{Get element names from list.}
\usage{
get_names(object)

\S4method{get_names}{pmx_list}(object)
}
\arguments{
\item{object}{list object}
}
\value{
character vector with all the element names of this list
}
\description{
Get element names from list.
}
