% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{get_nonmem_name}
\alias{get_nonmem_name}
\alias{get_nonmem_name,theta-method}
\alias{get_nonmem_name,omega-method}
\alias{get_nonmem_name,sigma-method}
\title{Get NONMEM name.}
\usage{
get_nonmem_name(object)

\S4method{get_nonmem_name}{theta}(object)

\S4method{get_nonmem_name}{omega}(object)

\S4method{get_nonmem_name}{sigma}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
the NONMEM name associated with this object
}
\description{
Get NONMEM name.
}
