% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_blocks.R
\name{get_omega_block}
\alias{get_omega_block}
\alias{get_omega_block,omega_blocks,double_array_parameter-method}
\title{Get the right block of OMEGA's.}
\usage{
get_omega_block(object, x)

\S4method{get_omega_block}{omega_blocks,double_array_parameter}(object, x)
}
\arguments{
\item{object}{list of OMEGA blocks}

\item{x}{omega param}
}
\value{
the corresponding OMEGA block or NULL if not found
}
\description{
Get the right block of OMEGA's.
}
