% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_block.R
\name{get_omega_indexes}
\alias{get_omega_indexes}
\alias{get_omega_indexes,omega_block-method}
\title{Get the indexes of the omegas.}
\usage{
get_omega_indexes(object)

\S4method{get_omega_indexes}{omega_block}(object)
}
\arguments{
\item{object}{omega block}
}
\value{
a list of integers
}
\description{
Get the indexes of the omegas.
}
