% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_property.R,
%   R/compartment_bioavailability.R, R/compartment_lag_time.R,
%   R/compartment_infusion_duration.R, R/compartment_infusion_rate.R,
%   R/compartment_initial_condition.R
\name{get_prefix}
\alias{get_prefix}
\alias{get_prefix,compartment_bioavailability-method}
\alias{get_prefix,compartment_lag_time-method}
\alias{get_prefix,compartment_infusion_duration-method}
\alias{get_prefix,compartment_infusion_rate-method}
\alias{get_prefix,compartment_initial_condition-method}
\title{Get prefix.}
\usage{
get_prefix(object, ...)

\S4method{get_prefix}{compartment_bioavailability}(object, ...)

\S4method{get_prefix}{compartment_lag_time}(object, ...)

\S4method{get_prefix}{compartment_infusion_duration}(object, ...)

\S4method{get_prefix}{compartment_infusion_rate}(object, ...)

\S4method{get_prefix}{compartment_initial_condition}(object, ...)
}
\arguments{
\item{object}{generic object}

\item{...}{e.g. \code{dest='mrgsolve'}}
}
\value{
the prefix of this object
}
\description{
Get prefix.
}
