% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_record_delimiter}
\alias{get_record_delimiter}
\title{Get record delimiter.}
\usage{
get_record_delimiter(line)
}
\arguments{
\item{line}{any line, single character value}
}
\value{
the record delimiter between brackets
}
\description{
Get record delimiter.
}
