% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_record_equation_names}
\alias{get_record_equation_names}
\title{Get record equation names}
\usage{
get_record_equation_names(record)
}
\arguments{
\item{record}{any code record}
}
\value{
a character vector with the equation names
}
\description{
Get record equation names
}
\keyword{internal}
