% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_property.R,
%   R/compartment_bioavailability.R, R/compartment_lag_time.R,
%   R/compartment_infusion_duration.R, R/compartment_infusion_rate.R,
%   R/compartment_initial_condition.R
\name{get_record_name}
\alias{get_record_name}
\alias{get_record_name,compartment_bioavailability-method}
\alias{get_record_name,compartment_lag_time-method}
\alias{get_record_name,compartment_infusion_duration-method}
\alias{get_record_name,compartment_infusion_rate-method}
\alias{get_record_name,compartment_initial_condition-method}
\title{Get record name.}
\usage{
get_record_name(object)

\S4method{get_record_name}{compartment_bioavailability}(object)

\S4method{get_record_name}{compartment_lag_time}(object)

\S4method{get_record_name}{compartment_infusion_duration}(object)

\S4method{get_record_name}{compartment_infusion_rate}(object)

\S4method{get_record_name}{compartment_initial_condition}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
the name of the record
}
\description{
Get record name.
}
