% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/parameter.R, R/parameters.R,
%   R/campsis_model.R
\name{get_uncertainty}
\alias{get_uncertainty}
\alias{get_uncertainty,parameter-method}
\alias{get_uncertainty,parameters-method}
\alias{get_uncertainty,campsis_model-method}
\title{Get uncertainty on the parameters.}
\usage{
get_uncertainty(object, ...)

\S4method{get_uncertainty}{parameter}(object, varcov, ...)

\S4method{get_uncertainty}{parameters}(object, ...)

\S4method{get_uncertainty}{campsis_model}(object, ...)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments}

\item{varcov}{variance covariance matrix}
}
\value{
data frame with standard error (se) and relative standard error (rse%) columns
}
\description{
Get uncertainty on the parameters.
}
