% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/parameters.R, R/campsis_model.R
\name{get_var_cov}
\alias{get_var_cov}
\alias{get_var_cov,parameters-method}
\alias{get_var_cov,campsis_model-method}
\title{Get variance-covariance matrix.}
\usage{
get_var_cov(object)

\S4method{get_var_cov}{parameters}(object)

\S4method{get_var_cov}{campsis_model}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
a variance-covariance matrix (data frame) or NULL if no matrix present
}
\description{
Get variance-covariance matrix.
}
