% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{getByName}
\alias{getByName}
\title{#' Get an element from a list by name.}
\usage{
getByName(object, name)
}
\arguments{
\item{object}{list object}

\item{name}{element name to search for}
}
\value{
the element that was found or NULL if no element was found with the same name
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`getByName()` is deprecated in favor of `get_by_name()`.
}
\keyword{internal}
