% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCampsismodOption}
\alias{getCampsismodOption}
\title{Get Campsismod option logic.}
\usage{
getCampsismodOption(name, default)
}
\arguments{
\item{name}{option to search}

\item{default}{default value if option not found}
}
\value{
option value
}
\description{
Get Campsismod option logic.
}
