% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCampsismodOptions}
\alias{getCampsismodOptions}
\title{Get the Campsismod options (R options).}
\usage{
getCampsismodOptions()
}
\value{
global options for Campsismod
}
\description{
Get the Campsismod options (R options).
}
\keyword{internal}
