% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{getCompartmentIndex}
\alias{getCompartmentIndex}
\title{Get the compartment index for the specified compartment name.}
\usage{
getCompartmentIndex(object, name)
}
\arguments{
\item{object}{generic object that contains compartments information}

\item{name}{compartment name}
}
\value{
the corresponding compartment index
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`getCompartmentIndex()` is deprecated in favor of `get_compartment_index()`.
}
\keyword{internal}
