% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{getCompartments}
\alias{getCompartments}
\title{Get all compartments from model.
Only for model instantiation. Not exported.}
\usage{
getCompartments(model)
}
\arguments{
\item{model}{model}
}
\value{
compartments object
}
\description{
Get all compartments from model.
Only for model instantiation. Not exported.
}
