% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{getEquation}
\alias{getEquation}
\title{Get equation from code record(s).}
\usage{
getEquation(object, lhs)
}
\arguments{
\item{object}{object containing code record(s)}

\item{lhs}{left-hand-side variable}
}
\value{
equation, character or NULL if not found
}
\description{
Get equation from code record(s).
}
