% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_record.R
\name{getEquationIndex}
\alias{getEquationIndex}
\title{Get equation index.}
\usage{
getEquationIndex(object, lhs)
}
\arguments{
\item{object}{code record}

\item{lhs}{left-hand-side variable to search for}
}
\value{
index in this record or -1 if not found
}
\description{
Get equation index.
}
