% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getInitialConditionCmt}
\alias{getInitialConditionCmt}
\title{Get initial condition compartment.
Assumes x is an initial condition (isInitialCondition already called).}
\usage{
getInitialConditionCmt(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical vector
}
\description{
Get initial condition compartment.
Assumes x is an initial condition (isInitialCondition already called).
}
