% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{getMappingMatrix}
\alias{getMappingMatrix}
\title{Return a matrix filled in with OMEGA/SIGMA names to be mapped with the values.
Un-existing parameters are filled in with the empty string.}
\usage{
getMappingMatrix(parameters, type)
}
\arguments{
\item{parameters}{subset of parameters}

\item{type}{type of parameter to map (omega or sigma)}
}
\value{
a matrix with the names of the OMEGA/SIGMA parameters
}
\description{
Return a matrix filled in with OMEGA/SIGMA names to be mapped with the values.
Un-existing parameters are filled in with the empty string.
}
