% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_element_list.R,
%   R/model_unknown_statement.R, R/model_comment.R, R/model_line_break.R,
%   R/model_equation.R, R/model_ode.R, R/model_if_statement.R, R/compartment.R,
%   R/compartment_bioavailability.R, R/compartment_lag_time.R,
%   R/compartment_infusion_duration.R, R/compartment_infusion_rate.R,
%   R/compartment_initial_condition.R, R/parameter.R, R/code_record.R,
%   R/omega_block.R
\name{getName}
\alias{getName}
\alias{getName,unknown_statement-method}
\alias{getName,comment-method}
\alias{getName,line_break-method}
\alias{getName,equation-method}
\alias{getName,ode-method}
\alias{getName,if_statement-method}
\alias{getName,compartment-method}
\alias{getName,compartment_bioavailability-method}
\alias{getName,compartment_lag_time-method}
\alias{getName,compartment_infusion_duration-method}
\alias{getName,compartment_infusion_rate-method}
\alias{getName,compartment_initial_condition-method}
\alias{getName,theta-method}
\alias{getName,omega-method}
\alias{getName,sigma-method}
\alias{getName,main_record-method}
\alias{getName,ode_record-method}
\alias{getName,f_record-method}
\alias{getName,lag_record-method}
\alias{getName,duration_record-method}
\alias{getName,rate_record-method}
\alias{getName,init_record-method}
\alias{getName,error_record-method}
\alias{getName,omega_block-method}
\title{Get element name.}
\usage{
getName(x)

\S4method{getName}{unknown_statement}(x)

\S4method{getName}{comment}(x)

\S4method{getName}{line_break}(x)

\S4method{getName}{equation}(x)

\S4method{getName}{ode}(x)

\S4method{getName}{if_statement}(x)

\S4method{getName}{compartment}(x)

\S4method{getName}{compartment_bioavailability}(x)

\S4method{getName}{compartment_lag_time}(x)

\S4method{getName}{compartment_infusion_duration}(x)

\S4method{getName}{compartment_infusion_rate}(x)

\S4method{getName}{compartment_initial_condition}(x)

\S4method{getName}{theta}(x)

\S4method{getName}{omega}(x)

\S4method{getName}{sigma}(x)

\S4method{getName}{main_record}(x)

\S4method{getName}{ode_record}(x)

\S4method{getName}{f_record}(x)

\S4method{getName}{lag_record}(x)

\S4method{getName}{duration_record}(x)

\S4method{getName}{rate_record}(x)

\S4method{getName}{init_record}(x)

\S4method{getName}{error_record}(x)

\S4method{getName}{omega_block}(x)
}
\arguments{
\item{x}{element to know the name}
}
\value{
the name of this element
}
\description{
Get element name.
}
