% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{getNameInModel}
\alias{getNameInModel}
\title{Get the name of the given parameter in the Campsis model.}
\usage{
getNameInModel(x)
}
\arguments{
\item{x}{element to know the name}
}
\value{
the name of this parameter
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`getNameInModel()` is deprecated in favor of `get_name_in_model()`.
}
\keyword{internal}
