% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{getNames}
\alias{getNames}
\title{Get element names from list.}
\usage{
getNames(object)
}
\arguments{
\item{object}{list object}
}
\value{
character vector with all the element names of this list
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`getNames()` is deprecated in favor of `get_names()`.
}
\keyword{internal}
