% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_blocks.R
\name{getOmegaBlock}
\alias{getOmegaBlock}
\alias{getOmegaBlock,omega_blocks,double_array_parameter-method}
\title{Get the right block of OMEGA's.}
\usage{
getOmegaBlock(object, x)

\S4method{getOmegaBlock}{omega_blocks,double_array_parameter}(object, x)
}
\arguments{
\item{object}{list of OMEGA blocks}

\item{x}{omega param}
}
\value{
the corresponding OMEGA block or NULL if not found
}
\description{
Get the right block of OMEGA's.
}
