% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_block.R
\name{getOmegaIndexes}
\alias{getOmegaIndexes}
\alias{getOmegaIndexes,omega_block-method}
\title{Get the indexes of the omegas.}
\usage{
getOmegaIndexes(object)

\S4method{getOmegaIndexes}{omega_block}(object)
}
\arguments{
\item{object}{omega block}
}
\value{
a list of integers
}
\description{
Get the indexes of the omegas.
}
