% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{getParameter}
\alias{getParameter}
\title{Get parameter function (single index).}
\usage{
getParameter(object, type, index, index2)

getParameter(object, type, index, index2)
}
\arguments{
\item{object}{generic object}

\item{type}{parameter type: omega or sigma}

\item{index}{first index}

\item{index2}{second index}
}
\value{
filtered object

filtered object
}
\description{
Get parameter function (single index).

Get parameter function (double index).
}
