% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_characteristic.R
\name{getPrefix}
\alias{getPrefix}
\title{Get prefix.}
\usage{
getPrefix(object, ...)
}
\arguments{
\item{object}{generic object}

\item{...}{e.g. dest='mrgsolve'}
}
\description{
Get prefix.
}
