% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{getRecord}
\alias{getRecord}
\title{Get record.}
\usage{
getRecord(object, recordType)
}
\arguments{
\item{object}{generic object}

\item{recordType}{parameter type: theta}
}
\value{
record object
}
\description{
Get record.
}
