% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getRecordDelimiter}
\alias{getRecordDelimiter}
\title{Get record delimiter.}
\usage{
getRecordDelimiter(line)
}
\arguments{
\item{line}{any line, single character value}
}
\value{
the record delimiter between brackets
}
\description{
Get record delimiter.
}
