% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getRecordEquationNames}
\alias{getRecordEquationNames}
\title{Get record equation names}
\usage{
getRecordEquationNames(record)
}
\arguments{
\item{record}{any code record}
}
\value{
a character vector with the equation names
}
\description{
Get record equation names
}
