% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_property.R,
%   R/compartment_bioavailability.R, R/compartment_lag_time.R,
%   R/compartment_infusion_duration.R, R/compartment_infusion_rate.R,
%   R/compartment_initial_condition.R
\name{getRecordName}
\alias{getRecordName}
\alias{getRecordName,compartment_bioavailability-method}
\alias{getRecordName,compartment_lag_time-method}
\alias{getRecordName,compartment_infusion_duration-method}
\alias{getRecordName,compartment_infusion_rate-method}
\alias{getRecordName,compartment_initial_condition-method}
\title{Get record name.}
\usage{
getRecordName(object)

\S4method{getRecordName}{compartment_bioavailability}(object)

\S4method{getRecordName}{compartment_lag_time}(object)

\S4method{getRecordName}{compartment_infusion_duration}(object)

\S4method{getRecordName}{compartment_infusion_rate}(object)

\S4method{getRecordName}{compartment_initial_condition}(object)
}
\arguments{
\item{object}{generic object}
}
\description{
Get record name.
}
