% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/parameters.R, R/campsis_model.R
\name{getVarCov}
\alias{getVarCov}
\alias{getVarCov,parameters-method}
\alias{getVarCov,campsis_model-method}
\title{Get variance-covariance matrix.}
\usage{
getVarCov(object)

\S4method{getVarCov}{parameters}(object)

\S4method{getVarCov}{campsis_model}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
a variance-covariance matrix (data frame) or NULL if no matrix present
}
\description{
Get variance-covariance matrix.
}
