% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_interface.R
\name{has_exact_method}
\alias{has_exact_method}
\title{Has exact method allows to check the existence of a S4 method in Campsis
based on its signature.}
\usage{
has_exact_method(generic, signature, where = topenv(parent.frame()))
}
\arguments{
\item{generic}{generic function name}

\item{signature}{function signature}

\item{where}{where to search functions}
}
\value{
logical value
}
\description{
Has exact method allows to check the existence of a S4 method in Campsis
based on its signature.
}
