% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_block.R
\name{has_off_diagonal_omegas}
\alias{has_off_diagonal_omegas}
\alias{has_off_diagonal_omegas,omega_block-method}
\title{Has off-diagonal omegas.}
\usage{
has_off_diagonal_omegas(object)

\S4method{has_off_diagonal_omegas}{omega_block}(object)
}
\arguments{
\item{object}{omega block}
}
\value{
TRUE or FALSE
}
\description{
Has off-diagonal omegas.
}
