% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hasComment}
\alias{hasComment}
\title{Check if string contains Campsis-style comments.}
\usage{
hasComment(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical value
}
\description{
Check if string contains Campsis-style comments.
}
