% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/code_record.R, R/code_records.R,
%   R/campsis_model.R
\name{hasEquation}
\alias{hasEquation}
\alias{hasEquation,code_record,character-method}
\alias{hasEquation,code_records,character-method}
\alias{hasEquation,campsis_model,character-method}
\title{Has equation.}
\usage{
hasEquation(object, lhs)

\S4method{hasEquation}{code_record,character}(object, lhs)

\S4method{hasEquation}{code_records,character}(object, lhs)

\S4method{hasEquation}{campsis_model,character}(object, lhs)
}
\arguments{
\item{object}{object containing code record(s)}

\item{lhs}{left-hand-side variable to search}
}
\value{
TRUE or FALSE
}
\description{
Has equation.
}
