% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_block.R
\name{hasOffDiagonalOmegas}
\alias{hasOffDiagonalOmegas}
\alias{hasOffDiagonalOmegas,omega_block-method}
\title{Has off-diagonal omegas.}
\usage{
hasOffDiagonalOmegas(object)

\S4method{hasOffDiagonalOmegas}{omega_block}(object)
}
\arguments{
\item{object}{omega block}
}
\value{
TRUE or FALSE
}
\description{
Has off-diagonal omegas.
}
