% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{hasParameter}
\alias{hasParameter}
\title{Check if list has given parameter.}
\usage{
hasParameter(object, parameter)
}
\arguments{
\item{object}{generic object}

\item{parameter}{parameter to check}
}
\value{
the parameter that matches
}
\description{
Check if list has given parameter.
}
