% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{identifyModelParametersFromVarcov}
\alias{identifyModelParametersFromVarcov}
\title{Identify the model parameters present in the variance-covariance matrix.}
\usage{
identifyModelParametersFromVarcov(parameters)
}
\arguments{
\item{parameters}{all Campsis parameters, including the variance-covariance matrix}
}
\value{
subset of Campsis parameters
}
\description{
Identify the model parameters present in the variance-covariance matrix.
}
