% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_if_statement.R
\docType{class}
\name{if_statement-class}
\alias{if_statement-class}
\title{If-statement class. Any statement in the form if (condition) A = B.}
\description{
If-statement class. Any statement in the form if (condition) A = B.
}
\section{Slots}{

\describe{
\item{\code{condition}}{IF statement condition}

\item{\code{equation}}{any equation or ODE}
}}

