% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{if_statement_pattern_str}
\alias{if_statement_pattern_str}
\title{Return the IF-statement pattern (string form).}
\usage{
if_statement_pattern_str()
}
\value{
pattern (regular expression)
}
\description{
Return the IF-statement pattern (string form).
}
\keyword{internal}
