% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_if_statement.R
\name{IfStatement}
\alias{IfStatement}
\title{Create a new IF-statement.}
\usage{
IfStatement(condition, equation, comment = as.character(NA))
}
\arguments{
\item{condition}{condition, single character string}

\item{equation}{equation if condition is met}

\item{comment}{comment if any, single character string}
}
\value{
an IF-statement
}
\description{
Create a new IF-statement.
}
