% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ifStatementPattern}
\alias{ifStatementPattern}
\title{Return the IF-statement pattern.}
\usage{
ifStatementPattern()
}
\value{
pattern (regular expression)
}
\description{
Return the IF-statement pattern.
}
\keyword{internal}
