% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ifStatementPatternStr}
\alias{ifStatementPatternStr}
\title{Return the IF-statement pattern (string form).}
\usage{
ifStatementPatternStr()
}
\value{
pattern (regular expression)
}
\description{
Return the IF-statement pattern (string form).
}
\keyword{internal}
