% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\name{index_of}
\alias{index_of}
\alias{index_of,pmx_list,pmx_element-method}
\title{Get the index of an element in list.}
\usage{
index_of(object, x)

\S4method{index_of}{pmx_list,pmx_element}(object, x)
}
\arguments{
\item{object}{list object}

\item{x}{element to know the index}
}
\value{
index of this element
}
\description{
Get the index of an element in list.
}
