% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_infusion_duration.R
\name{InfusionDuration}
\alias{InfusionDuration}
\title{Create an infusion duration/rate.}
\usage{
InfusionDuration(compartment, rhs, rate = FALSE)
}
\arguments{
\item{compartment}{compartment index}

\item{rhs}{right-hand side part of the equation}

\item{rate}{logical value, TRUE if distribution is a rate, FALSE if it is an infusion}
}
\value{
details about infusion duration/rate
}
\description{
Create an infusion duration/rate.
}
