% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_infusion_rate.R
\name{InfusionRate}
\alias{InfusionRate}
\title{Create an infusion rate.}
\usage{
InfusionRate(compartment, rhs = "")
}
\arguments{
\item{compartment}{compartment index}

\item{rhs}{right-hand side part of the equation}
}
\value{
details about infusion rate
}
\description{
Create an infusion rate.
}
