% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_initial_condition.R
\name{InitialCondition}
\alias{InitialCondition}
\title{Create an initial condition.}
\usage{
InitialCondition(compartment, rhs)
}
\arguments{
\item{compartment}{compartment index}

\item{rhs}{right-hand side part of the equation}
}
\value{
initial condition object
}
\description{
Create an initial condition.
}
