% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_comment}
\alias{is_comment}
\title{Check if string is a Campsis comment (i.e. not an equation).}
\usage{
is_comment(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical value
}
\description{
Check if string is a Campsis comment (i.e. not an equation).
}
