% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{is_diag}
\alias{is_diag}
\alias{is_diag,double_array_parameter-method}
\title{Is diagonal.}
\usage{
is_diag(object)

\S4method{is_diag}{double_array_parameter}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
logical value
}
\description{
Is diagonal.
}
