% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_empty_line}
\alias{is_empty_line}
\title{Check if string is an empty line.}
\usage{
is_empty_line(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical value
}
\description{
Check if string is an empty line.
}
