% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_equation}
\alias{is_equation}
\title{Say if line in record is an equation not.}
\usage{
is_equation(x)
}
\arguments{
\item{x}{character value}
}
\value{
logical value
}
\description{
Say if line in record is an equation not.
}
