% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_if_statement}
\alias{is_if_statement}
\title{Say if line in record is an IF-statement.}
\usage{
is_if_statement(x)
}
\arguments{
\item{x}{character value}
}
\value{
logical value
}
\description{
Say if line in record is an IF-statement.
}
