% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_record_delimiter}
\alias{is_record_delimiter}
\title{Is record delimiter. A record delimiter is any line starting with [...].}
\usage{
is_record_delimiter(line)
}
\arguments{
\item{line}{any line, single character value}
}
\value{
a logical value
}
\description{
Is record delimiter. A record delimiter is any line starting with [...].
}
