% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_rxode}
\alias{is_rxode}
\title{Check if the destination engine is RxODE or rxode2.
Note that rxode2 is the successor of RxODE.}
\usage{
is_rxode(dest)
}
\arguments{
\item{dest}{destination engine}
}
\value{
TRUE if RxODE or rxode2, FALSE otherwise
}
\description{
Check if the destination engine is RxODE or rxode2.
Note that rxode2 is the successor of RxODE.
}
\keyword{internal}
