% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isBioavailibility}
\alias{isBioavailibility}
\title{Say if line(s) in record is/are bioavailabilities.}
\usage{
isBioavailibility(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical vector
}
\description{
Say if line(s) in record is/are bioavailabilities.
}
