% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{isBlockFixed}
\alias{isBlockFixed}
\title{Say if the block is fixed (i.e. all parameters are fixed).}
\usage{
isBlockFixed(block)
}
\arguments{
\item{block}{block of OMEGAs or SIGMAs}
}
\value{
logical value
}
\description{
Say if the block is fixed (i.e. all parameters are fixed).
}
