% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{isDiag}
\alias{isDiag}
\title{Is diagonal.}
\usage{
isDiag(object)
}
\arguments{
\item{object}{generic object}
}
\value{
logical value
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`isDiag()` is deprecated in favor of `is_diag()`.
}
