% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isEmptyLine}
\alias{isEmptyLine}
\title{Check if string is an empty line.}
\usage{
isEmptyLine(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical value
}
\description{
Check if string is an empty line.
}
