% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isIfStatement}
\alias{isIfStatement}
\title{Say if line in record is an IF-statement.}
\usage{
isIfStatement(x)
}
\arguments{
\item{x}{character value}
}
\value{
logical value
}
\description{
Say if line in record is an IF-statement.
}
