% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isInfusionDuration}
\alias{isInfusionDuration}
\title{Say if line(s) in record is/are infusion durations.}
\usage{
isInfusionDuration(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical vector
}
\description{
Say if line(s) in record is/are infusion durations.
}
