% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isInitialCondition}
\alias{isInitialCondition}
\title{Say if line(s) in record is/are initial conditions.}
\usage{
isInitialCondition(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical vector
}
\description{
Say if line(s) in record is/are initial conditions.
}
