% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isLagTime}
\alias{isLagTime}
\title{Say if line(s) in record is/are lag times.}
\usage{
isLagTime(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical vector
}
\description{
Say if line(s) in record is/are lag times.
}
