% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{isMatrixPositiveDefinite}
\alias{isMatrixPositiveDefinite}
\title{Is matrix positive definite. Same check as \code{mvtnorm} does.}
\usage{
isMatrixPositiveDefinite(matrix, tol = 1e-06)
}
\arguments{
\item{matrix}{matrix to check}

\item{tol}{tolerance when checking the eigenvalues}
}
\description{
Is matrix positive definite. Same check as \code{mvtnorm} does.
}
