% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isODE}
\alias{isODE}
\title{Say if line(s) in record is/are ODE or not.}
\usage{
isODE(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical vector
}
\description{
Say if line(s) in record is/are ODE or not.
}
