% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isRate}
\alias{isRate}
\title{Say if line(s) in record is/are rates.}
\usage{
isRate(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical vector
}
\description{
Say if line(s) in record is/are rates.
}
