% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_interface.R
\name{json_to_parameter}
\alias{json_to_parameter}
\title{JSON to Campsis parameter.}
\usage{
json_to_parameter(x, index = NULL, index2 = NULL)
}
\arguments{
\item{x}{JSON data}

\item{index}{parameter index to add}

\item{index2}{second parameter index to add for OMEGAs and SIGMAs}
}
\value{
Campsis parameter
}
\description{
JSON to Campsis parameter.
}
