% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_element.R
\name{JSONElement}
\alias{JSONElement}
\title{Instantiate a JSON element.}
\usage{
JSONElement(x)
}
\arguments{
\item{x}{JSON object representation in R}
}
\value{
JSON element
}
\description{
Instantiate a JSON element.
}
