% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_interface.R
\name{jsonToOffDiagParameter}
\alias{jsonToOffDiagParameter}
\title{Convert JSON correlation parameter (OMEGA or SIGMA) into a Campsis parameter.}
\usage{
jsonToOffDiagParameter(json, diag_names)
}
\arguments{
\item{json}{JSON data}

\item{diag_names}{parameter names on the diagonal, character vector}
}
\value{
the corresponding Campsis parameter
}
\description{
Convert JSON correlation parameter (OMEGA or SIGMA) into a Campsis parameter.
}
