% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{jsonToParameter}
\alias{jsonToParameter}
\title{JSON to Campsis parameter.}
\usage{
jsonToParameter(x, index = NULL, index2 = NULL)
}
\arguments{
\item{x}{JSON data}

\item{index}{parameter index to add}

\item{index2}{second parameter index to add for OMEGAs and SIGMAs}
}
\value{
Campsis parameter
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`jsonToParameter()` is deprecated in favor of `json_to_parameter()`.
}
\keyword{internal}
