% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\name{keep}
\alias{keep}
\alias{keep,ANY-method}
\title{Keep element(s) from list using predicate function.}
\usage{
keep(.x, .p, ...)

\S4method{keep}{ANY}(.x, .p, ...)
}
\arguments{
\item{.x}{a list, a vector or a PMX list}

\item{.p}{a predicate function}

\item{...}{extra arguments, ignored}
}
\value{
updated list, vector or PMX list
}
\description{
Keep element(s) from list using predicate function.
}
