% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_block.R
\name{length,omega_block-method}
\alias{length,omega_block-method}
\title{Return the number of OMEGA's on the diagonal.}
\usage{
\S4method{length}{omega_block}(x)
}
\arguments{
\item{x}{omega block}
}
\value{
a number
}
\description{
Return the number of OMEGA's on the diagonal.
}
