% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R, R/code_record.R,
%   R/omega_block.R
\name{length,pmx_list-method}
\alias{length,pmx_list-method}
\alias{length,statements_record-method}
\alias{length,omega_block-method}
\title{Return the length of this list.}
\usage{
\S4method{length}{pmx_list}(x)

\S4method{length}{statements_record}(x)

\S4method{length}{omega_block}(x)
}
\arguments{
\item{x}{omega block}
}
\value{
the length of this list, integer value

a number
}
\description{
Return the length of this list.

Return the number of OMEGA's on the diagonal.
}
\keyword{internal}
