% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/parameter.R, R/campsis_model.R
\name{load_from_json}
\alias{load_from_json}
\alias{load_from_json,theta,json_element-method}
\alias{load_from_json,omega,json_element-method}
\alias{load_from_json,sigma,json_element-method}
\alias{load_from_json,campsis_model,json_element-method}
\alias{load_from_json,campsis_model,character-method}
\title{Fill-in S4 object from the JSON content.}
\usage{
load_from_json(object, json)

\S4method{load_from_json}{theta,json_element}(object, json)

\S4method{load_from_json}{omega,json_element}(object, json)

\S4method{load_from_json}{sigma,json_element}(object, json)

\S4method{load_from_json}{campsis_model,json_element}(object, json)

\S4method{load_from_json}{campsis_model,character}(object, json)
}
\arguments{
\item{object}{pre-initiated S4 object}

\item{json}{JSON (usually a list)}
}
\value{
the loaded S4 object
}
\description{
Fill-in S4 object from the JSON content.
}
