% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{loadFromJSON}
\alias{loadFromJSON}
\title{Fill-in S4 object from the JSON content.}
\usage{
loadFromJSON(object, json)
}
\arguments{
\item{object}{pre-initiated S4 object}

\item{json}{JSON (usually a list)}
}
\value{
the loaded S4 object
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`loadFromJSON()` is deprecated in favor of `load_from_json()`.
}
