% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_settings.R
\docType{class}
\name{manual_replication_settings-class}
\alias{manual_replication_settings-class}
\title{Manual replication settings class.}
\description{
Manual replication settings class.
}
\section{Slots}{

\describe{
\item{\code{replicated_parameters}}{data frame, 1 row per replicate}
}}

