% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_interface.R
\name{map_json_properties_to_s4_slots}
\alias{map_json_properties_to_s4_slots}
\title{Map JSON properties to S4 slots.}
\usage{
map_json_properties_to_s4_slots(object, json, discard_type = TRUE)
}
\arguments{
\item{object}{S4 object}

\item{json}{JSON element, json_element class}

\item{discard_type}{discard JSON property 'type'}
}
\value{
a S4 object
}
\description{
Map JSON properties to S4 slots.
}
