% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_interface.R
\name{map_s4_slots_to_json_properties}
\alias{map_s4_slots_to_json_properties}
\title{Map S4 slots to JSON properties.}
\usage{
map_s4_slots_to_json_properties(
  object,
  add_type = TRUE,
  optional = NULL,
  ignore = NULL
)
}
\arguments{
\item{object}{S4 object}

\item{add_type}{add type as a property, TRUE by default}

\item{optional}{properties that are optional in JSON, character vector}

\item{ignore}{slots to be ignored}
}
\value{
a JSON object ready to be serialised
}
\description{
Map S4 slots to JSON properties.
}
