% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{mapJSONPropertiesToS4Slots}
\alias{mapJSONPropertiesToS4Slots}
\title{Map JSON properties to S4 slots.}
\usage{
mapJSONPropertiesToS4Slots(object, json, discard_type = TRUE)
}
\value{
a JSON object ready to be serialised
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`mapJSONPropertiesToS4Slots()` is deprecated in favor of `map_json_properties_to_s4_slots()`.
}
\keyword{internal}
