% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{match_double_array_parameter}
\alias{match_double_array_parameter}
\title{Match double array parameter from list based on the name instead of the index.
If a match is found, its indexes are automatically copied.}
\usage{
match_double_array_parameter(object, x)
}
\arguments{
\item{object}{parameters}

\item{x}{double array parameter to match}
}
\value{
the same parameter is no match was found or the same parameter with updated indexes if a match was found
}
\description{
Match double array parameter from list based on the name instead of the index.
If a match is found, its indexes are automatically copied.
}
\keyword{internal}
