% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{match_single_array_parameter}
\alias{match_single_array_parameter}
\title{Match single array parameter from list based on the name instead of the index.
If a match is found, its index is automatically copied.}
\usage{
match_single_array_parameter(object, x)
}
\arguments{
\item{object}{parameters}

\item{x}{single array parameter to match}
}
\value{
the same parameter is no match was found or the same parameter with updated index if a match was found
}
\description{
Match single array parameter from list based on the name instead of the index.
If a match is found, its index is automatically copied.
}
\keyword{internal}
