% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{max_index}
\alias{max_index}
\alias{max_index,parameters-method}
\title{Max index.}
\usage{
max_index(object)

\S4method{max_index}{parameters}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
max index
}
\description{
Max index.
}
