% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{maxIndex}
\alias{maxIndex}
\title{Max index.}
\usage{
maxIndex(object)
}
\arguments{
\item{object}{generic object}
}
\value{
max index
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`maxIndex()` is deprecated in favor of `max_index()`.
}
\keyword{internal}
