% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{methods_global_functions}
\alias{methods_global_functions}
\title{A empty method only used to declare some of the global function definitions
contained in the package 'methods', which are used intensively.}
\usage{
methods_global_functions()
}
\value{
no return value
}
\description{
A empty method only used to declare some of the global function definitions
contained in the package 'methods', which are used intensively.
}
\keyword{internal}
