% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{min_index}
\alias{min_index}
\alias{min_index,parameters-method}
\title{Min index.}
\usage{
min_index(object)

\S4method{min_index}{parameters}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
min index
}
\description{
Min index.
}
