% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{minMaxDefault}
\alias{minMaxDefault}
\title{Min/max default values for the given parameter.}
\usage{
minMaxDefault(parameter)
}
\arguments{
\item{parameter}{Campsis parameter}
}
\value{
a tibble with the min and max values
}
\description{
Min/max default values for the given parameter.
}
\keyword{internal}
