% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_statement.R
\docType{class}
\name{model_statement-class}
\alias{model_statement-class}
\title{Model statement class. Any statement in a code record.}
\description{
Model statement class. Any statement in a code record.
}
\section{Slots}{

\describe{
\item{\code{comment}}{a comment associated to this model statement}
}}

