% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_suite}
\alias{model_suite}
\title{Campsis model suite.}
\format{
A list with all the models:
\describe{
   \item{pk}{extensive list of pharmacokinetic (PK) model templates}
   \item{pd}{list of pharmacodynamic (PD) model templates, to be plugged into any pharmacokinetic (PK) model}
   \item{tmdd}{extensive list of target-mediated drug disposition (TMDD) model templates}
   \item{nonmem}{list of model templates translated from standard NONMEM control streams}
   \item{literature}{a couple of models coming from the literature}
   \item{other}{a couple of custom models}
}
}
\source{
\url{https://calvagone.github.io/campsis.doc/}

\url{https://www.iconplc.com/solutions/technologies/nonmem/}
}
\usage{
model_suite
}
\description{
A library of models of all kinds, ready to be simulated in Campsis. These model
templates are sorted into the following categories: pharmacokinetic (PK), pharmacodynamic (PD),
target-mediated drug disposition (TMDD), NONMEM, literature  and other (custom models).
}
\keyword{datasets}
