% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_statements.R
\name{ModelStatements}
\alias{ModelStatements}
\title{Create an empty list of model statements.}
\usage{
ModelStatements()
}
\value{
a model statements object
}
\description{
Create an empty list of model statements.
}
