% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/code_records.R, R/campsis_model.R
\name{move}
\alias{move}
\alias{move,code_records,model_statement,pmx_position-method}
\alias{move,code_records,list,pmx_position-method}
\alias{move,campsis_model,ANY,pmx_position-method}
\title{Move element 'x' from object to a certain place.}
\usage{
move(object, x, to, ...)

\S4method{move}{code_records,model_statement,pmx_position}(object, x, to, ...)

\S4method{move}{code_records,list,pmx_position}(object, x, to, ...)

\S4method{move}{campsis_model,ANY,pmx_position}(object, x, to, ...)
}
\arguments{
\item{object}{generic object (e.g. model, code records, etc.)}

\item{x}{element to move}

\item{to}{destination (e.g. a position)}

\item{...}{extra arguments, unused}
}
\value{
updated object
}
\description{
Move element 'x' from object to a certain place.
}
