% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{mrgsolve_block}
\alias{mrgsolve_block}
\title{Convert code record for mrgsolve.}
\usage{
mrgsolve_block(record, init = NULL, capture = FALSE)
}
\arguments{
\item{record}{code record}

\item{init}{name of mrgsolve block}

\item{capture}{'capture' instead of 'double'}
}
\value{
translated record for mrgsolve
}
\description{
Convert code record for mrgsolve.
}
