% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{mrgsolve_capture}
\alias{mrgsolve_capture}
\title{Get the CAPTURE block for mrgsolve.}
\usage{
mrgsolve_capture(outvars, model)
}
\arguments{
\item{outvars}{outvars in method simulate}

\item{model}{Campsis model}
}
\value{
CAPTURE block or character(0) if no variable in outvars
}
\description{
Get the CAPTURE block for mrgsolve.
}
