% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{mrgsolve_compartment}
\alias{mrgsolve_compartment}
\title{Get the compartment block for mrgsolve.}
\usage{
mrgsolve_compartment(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
character vector, each value is a line
}
\description{
Get the compartment block for mrgsolve.
}
