% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{mrgsolve_main}
\alias{mrgsolve_main}
\title{Get the MAIN block for mrgsolve.}
\usage{
mrgsolve_main(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
MAIN block
}
\description{
Get the MAIN block for mrgsolve.
}
