% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{mrgsolve_ode}
\alias{mrgsolve_ode}
\title{Get the ODE block for mrgsolve.}
\usage{
mrgsolve_ode(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
ODE block
}
\description{
Get the ODE block for mrgsolve.
}
