% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{mrgsolve_table}
\alias{mrgsolve_table}
\title{Get the TABLE block for mrgsolve.}
\usage{
mrgsolve_table(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
TABLE block if at least one line in error record, character(0) otherwise
}
\description{
Get the TABLE block for mrgsolve.
}
