% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{mrgsolveCapture}
\alias{mrgsolveCapture}
\title{Get the CAPTURE block for mrgsolve.}
\usage{
mrgsolveCapture(outvars, model)
}
\arguments{
\item{outvars}{outvars in method simulate}

\item{model}{CAMPSIS model}
}
\value{
CAPTURE block or character(0) if no variable in outvars
}
\description{
Get the CAPTURE block for mrgsolve.
}
