% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{mrgsolveCompartment}
\alias{mrgsolveCompartment}
\title{Get the compartment block for mrgsolve.}
\usage{
mrgsolveCompartment(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
character vector, each value is a line
}
\description{
Get the compartment block for mrgsolve.
}
