% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{mrgsolveMain}
\alias{mrgsolveMain}
\title{Get the MAIN block for mrgsolve.}
\usage{
mrgsolveMain(model)
}
\arguments{
\item{model}{CAMPSIS model}
}
\value{
MAIN block
}
\description{
Get the MAIN block for mrgsolve.
}
