% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{mrgsolveMatrix}
\alias{mrgsolveMatrix}
\title{Get the OMEGA/SIGMA matrix for mrgsolve.}
\usage{
mrgsolveMatrix(model, type = "omega")
}
\arguments{
\item{model}{Campsis model}

\item{type}{either omega or sigma}
}
\value{
named matrix or character(0) if matrix is empty
}
\description{
Get the OMEGA/SIGMA matrix for mrgsolve.
}
