% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{mrgsolveOde}
\alias{mrgsolveOde}
\title{Get the ODE block for mrgsolve.}
\usage{
mrgsolveOde(model)
}
\arguments{
\item{model}{CAMPSIS model}
}
\value{
ODE block
}
\description{
Get the ODE block for mrgsolve.
}
