% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_record.R
\name{OdeRecord}
\alias{OdeRecord}
\title{Create ODE code record.}
\usage{
OdeRecord(code = character())
}
\arguments{
\item{code}{code record}
}
\value{
an ODE code record
}
\description{
Create ODE code record.
}
