% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\docType{class}
\name{omega-class}
\alias{omega-class}
\title{Omega parameter class.}
\description{
Omega parameter class.
}
\section{Slots}{

\describe{
\item{\code{same}}{logical value, tell if this omega is the same as the previous one}
}}

