% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{Omega}
\alias{Omega}
\title{Create an OMEGA parameter.}
\usage{
Omega(name = NA, index, index2, value = NA, fix = FALSE)
}
\arguments{
\item{name}{parameter name, e.g. CL (prefix OMEGA will be added automatically)}

\item{index}{parameter index}

\item{index2}{second parameter index}

\item{value}{parameter value}

\item{fix}{parameter was fixed in estimation, logical value}
}
\value{
an OMEGA parameter
}
\description{
Create an OMEGA parameter.
}
