% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_interface.R
\name{openJSON}
\alias{openJSON}
\title{Open JSON file.}
\usage{
openJSON(json, schema = NULL)
}
\arguments{
\item{json}{JSON in its string form or path to JSON file}

\item{schema}{JSON schema}
}
\value{
parsed JSON object
}
\description{
Open JSON file.
}
\keyword{internal}
