% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\docType{class}
\name{parameter-class}
\alias{parameter-class}
\title{Parameter class. Any parameter in a pharmacometric model.}
\description{
Parameter class. Any parameter in a pharmacometric model.
}
\section{Slots}{

\describe{
\item{\code{name}}{parameter name, optional (although recommended)}

\item{\code{index}}{parameter index, integer}

\item{\code{value}}{parameter value (e.g. the estimated value from a modelling tool)}

\item{\code{fix}}{logical value, say if parameter was fixed in the modelling phase}

\item{\code{label}}{parameter label, any string}

\item{\code{comment}}{any comment on this parameter, any string}
}}

