% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{class}
\name{parameters-class}
\alias{parameters-class}
\title{Parameters class.}
\description{
Parameters class.
}
\section{Slots}{

\describe{
\item{\code{varcov}}{attached variance-covariance matrix for these parameters}
}}

