% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parser.R
\name{parse_if_statement}
\alias{parse_if_statement}
\title{Parse IF-statement.
Assumption: \code{is_if_statement} method already called and returned TRUE.}
\usage{
parse_if_statement(line, comment = as.character(NA))
}
\arguments{
\item{line}{IF-statement as single character string value, comment omitted}

\item{comment}{any comment, NA by default}
}
\value{
an IF statement object
}
\description{
Parse IF-statement.
Assumption: \code{is_if_statement} method already called and returned TRUE.
}
