% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parser.R
\name{parse_statements}
\alias{parse_statements}
\title{Parse statements code and return Campsis statements.}
\usage{
parse_statements(code)
}
\arguments{
\item{code}{character vector containing all statements (text form)}
}
\value{
a list of Campsis statements
}
\description{
Parse statements code and return Campsis statements.
}
