% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parser.R
\name{parseProperties}
\alias{parseProperties}
\title{Parse properties and return CAMPSIS statements.}
\usage{
parseProperties(code)
}
\arguments{
\item{code}{character vector containing all properties (text form)}
}
\value{
a list of equations
}
\description{
Parse properties and return CAMPSIS statements.
}
