% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/model_parser.R
\name{parseRecord}
\alias{parseRecord}
\alias{parseRecord,statement_record-method}
\alias{parseRecord,code_record-method}
\title{Parse code record.}
\usage{
parseRecord(object, ...)

\S4method{parseRecord}{statement_record}(object)

\S4method{parseRecord}{code_record}(object)
}
\arguments{
\item{object}{object containing code to be parsed}

\item{...}{extra arguments}
}
\value{
updated code record
}
\description{
Parse code record.
}
