% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parser.R
\name{parseStatements}
\alias{parseStatements}
\title{Parse statements code and return CAMPSIS statements.}
\usage{
parseStatements(code)
}
\arguments{
\item{code}{character vector containing all statements (text form)}
}
\value{
a list of CAMPSIS statements
}
\description{
Parse statements code and return CAMPSIS statements.
}
