% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\docType{class}
\name{pmx_list-class}
\alias{pmx_list-class}
\title{PMX list class.}
\description{
PMX list class.
}
\section{Slots}{

\describe{
\item{\code{list}}{effective list which will contain the elements}

\item{\code{type}}{type of the elements this list may contain}
}}

