% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_element_position.R
\name{Position}
\alias{Position}
\title{Element position in list.}
\usage{
Position(x, after = TRUE)
}
\arguments{
\item{x}{either an integer position or an list element}

\item{after}{element to be added will be added after x (if after is TRUE) or before x (if after is FALSE)}
}
\description{
Element position in list.
}
