% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{process_extra_arg}
\alias{process_extra_arg}
\title{Process extra arguments.}
\usage{
process_extra_arg(args, name, default = NULL, mandatory = FALSE)
}
\arguments{
\item{args}{arguments list}

\item{name}{argument name to retrieve}

\item{default}{default value if argument is not present}

\item{mandatory}{mandatory argument, logical value}
}
\value{
requested argument value
}
\description{
Process extra arguments.
}
