% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_interface.R
\name{process_json_double_array_parameter}
\alias{process_json_double_array_parameter}
\title{Process JSON double array parameter.}
\usage{
process_json_double_array_parameter(x)
}
\arguments{
\item{x}{JSON data, OMEGA or SIGMA parameter}
}
\value{
updated JSON data with updated 'name' field and removed 'name2' field
}
\description{
Process JSON double array parameter.
}
