% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{processDoubleArrayArguments}
\alias{processDoubleArrayArguments}
\title{Process double array arguments.}
\usage{
processDoubleArrayArguments(index, index2, type)
}
\arguments{
\item{index}{parameter index}

\item{index2}{second parameter index}

\item{type}{variance type}
}
\value{
variance type
}
\description{
Process double array arguments.
}
\keyword{internal}
