% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{processExtraArg}
\alias{processExtraArg}
\title{Process extra arguments.}
\usage{
processExtraArg(args, name, default = NULL, mandatory = FALSE)
}
\arguments{
\item{args}{arguments list}

\item{name}{argument name to retrieve}

\item{default}{default value if argument is not present}

\item{mandatory}{mandatory argument, logical value}
}
\value{
requested argument value
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`processExtraArg()` is deprecated in favor of `process_extra_arg()`.
}
\keyword{internal}
