% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{read.allparameters}
\alias{read.allparameters}
\title{Read all parameters files at once.}
\usage{
read.allparameters(folder)
}
\arguments{
\item{folder}{path to folder or path to zipped project}

\item{type}{parameter type: 'theta', 'omega' or 'sigma'}
}
\value{
a PMX model
}
\description{
Read all parameters files at once.
}
