% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campsis_model.R
\name{read.campsis}
\alias{read.campsis}
\title{Read a Campsis model.}
\usage{
read.campsis(file)
}
\arguments{
\item{file}{path to folder (old format) or path to JSON file (new format)}
}
\value{
Campsis model
}
\description{
Read a Campsis model.
}
