% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{read.model}
\alias{read.model}
\title{Read model file.}
\usage{
read.model(file = NULL, text = NULL)
}
\arguments{
\item{file}{path to file 'model.campsis'}

\item{text}{model file as text, character (single or multiple lines)}
}
\value{
records object
}
\description{
Read model file.
}
