% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_model.R
\name{read.parameter}
\alias{read.parameter}
\title{Read parameter file.}
\usage{
read.parameter(file, type)
}
\arguments{
\item{file}{path to folder or path to zipped project}

\item{type}{parameter type: 'theta', 'omega' or 'sigma'}
}
\value{
a PMX model
}
\description{
Read parameter file.
}
