% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{read.parameters}
\alias{read.parameters}
\title{Read parameters file.}
\usage{
read.parameters(file, type)
}
\arguments{
\item{file}{path to CSV file}

\item{type}{parameter type: 'theta', 'omega' or 'sigma'}
}
\value{
parameters sub list
}
\description{
Read parameters file.
}
