% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_model.R
\name{read.pmxmod}
\alias{read.pmxmod}
\title{Read PMX model file.}
\usage{
read.pmxmod(file)
}
\arguments{
\item{file}{path to folder or path to zipped project}
}
\value{
a PMX model
}
\description{
Read PMX model file.
}
