% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{read}
\alias{read}
\title{Generic read method to read data from a file or a folder.}
\usage{
read(file, ...)
}
\arguments{
\item{file}{path to the file or folder to be read}

\item{...}{extra arguments}
}
\value{
the object representation of the data contained in the file
}
\description{
Generic read method to read data from a file or a folder.
}
