% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{read.varcov}
\alias{read.varcov}
\title{Read variance-covariance file.}
\usage{
read.varcov(file)
}
\arguments{
\item{file}{path to CSV file}
}
\value{
variance-covariance matrix
}
\description{
Read variance-covariance file.
}
