% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_na_column}
\alias{remove_na_column}
\title{Remove given column(s) if it has only NA's.}
\usage{
remove_na_column(x, column)
}
\arguments{
\item{x}{any data frame}

\item{column}{column name(s)}
}
\value{
updated data frame
}
\description{
Remove given column(s) if it has only NA's.
}
\keyword{internal}
