% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{remove_trailing_line_breaks}
\alias{remove_trailing_line_breaks}
\title{Remove all trailing line breaks.}
\usage{
remove_trailing_line_breaks(x)
}
\arguments{
\item{x}{character vector}
}
\value{
a character vector
}
\description{
Remove all trailing line breaks.
}
\keyword{internal}
