% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{removeEquation}
\alias{removeEquation}
\title{Remove equation from code record.}
\usage{
removeEquation(object, lhs)
}
\arguments{
\item{object}{object containing code record(s)}

\item{lhs}{left-hand-side variable}
}
\value{
updated code record
}
\description{
Remove equation from code record.
}
