% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{removeNaColumn}
\alias{removeNaColumn}
\title{Remove given column(s) if it has only NA's.}
\usage{
removeNaColumn(x, column)
}
\arguments{
\item{x}{any data frame}

\item{column}{column name(s)}
}
\value{
updated data frame
}
\description{
Remove given column(s) if it has only NA's.
}
