% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{removeTrailingLineBreaks}
\alias{removeTrailingLineBreaks}
\title{Remove all trailing line breaks.}
\usage{
removeTrailingLineBreaks(x)
}
\arguments{
\item{x}{character vector}
}
\value{
a character vector
}
\description{
Remove all trailing line breaks.
}
\keyword{internal}
