% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R, R/compartments.R,
%   R/parameters.R, R/code_record.R, R/code_records.R, R/campsis_model.R
\name{replace}
\alias{replace}
\alias{replace,pmx_list,pmx_element-method}
\alias{replace,compartments,compartment_property-method}
\alias{replace,parameters,single_array_parameter-method}
\alias{replace,parameters,double_array_parameter-method}
\alias{replace,statements_record,model_statement-method}
\alias{replace,code_records,model_statement-method}
\alias{replace,campsis_model,compartment_property-method}
\alias{replace,campsis_model,parameter-method}
\alias{replace,campsis_model,code_record-method}
\alias{replace,campsis_model,model_statement-method}
\title{Replace element by another in list.}
\usage{
replace(object, x)

\S4method{replace}{pmx_list,pmx_element}(object, x)

\S4method{replace}{compartments,compartment_property}(object, x)

\S4method{replace}{parameters,single_array_parameter}(object, x)

\S4method{replace}{parameters,double_array_parameter}(object, x)

\S4method{replace}{statements_record,model_statement}(object, x)

\S4method{replace}{code_records,model_statement}(object, x)

\S4method{replace}{campsis_model,compartment_property}(object, x)

\S4method{replace}{campsis_model,parameter}(object, x)

\S4method{replace}{campsis_model,code_record}(object, x)

\S4method{replace}{campsis_model,model_statement}(object, x)
}
\arguments{
\item{object}{list object}

\item{x}{element to replace}
}
\value{
list object or an error if the element does not exist in the list
}
\description{
Replace element by another in list.
}
