% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/pattern.R, R/model_statement.R,
%   R/model_unknown_statement.R, R/model_equation.R, R/model_if_statement.R,
%   R/compartment_property.R, R/code_record.R, R/code_records.R,
%   R/campsis_model.R
\name{replace_all}
\alias{replace_all}
\alias{replace_all,character,variable_pattern,character-method}
\alias{replace_all,character,pattern,character-method}
\alias{replace_all,model_statement,pattern,character-method}
\alias{replace_all,unknown_statement,pattern,character-method}
\alias{replace_all,equation,pattern,character-method}
\alias{replace_all,if_statement,pattern,character-method}
\alias{replace_all,compartment_property,pattern,character-method}
\alias{replace_all,code_record,pattern,character-method}
\alias{replace_all,code_records,pattern,character-method}
\alias{replace_all,campsis_model,pattern,character-method}
\alias{replace_all,campsis_model,character,character-method}
\title{Replace all occurrences in object.}
\usage{
replace_all(object, pattern, replacement, ...)

\S4method{replace_all}{character,variable_pattern,character}(object, pattern, replacement, ...)

\S4method{replace_all}{character,pattern,character}(object, pattern, replacement, ...)

\S4method{replace_all}{model_statement,pattern,character}(object, pattern, replacement, ...)

\S4method{replace_all}{unknown_statement,pattern,character}(object, pattern, replacement, ...)

\S4method{replace_all}{equation,pattern,character}(object, pattern, replacement, ...)

\S4method{replace_all}{if_statement,pattern,character}(object, pattern, replacement, ...)

\S4method{replace_all}{compartment_property,pattern,character}(object, pattern, replacement, ...)

\S4method{replace_all}{code_record,pattern,character}(object, pattern, replacement, ...)

\S4method{replace_all}{code_records,pattern,character}(object, pattern, replacement, ...)

\S4method{replace_all}{campsis_model,pattern,character}(object, pattern, replacement, ...)

\S4method{replace_all}{campsis_model,character,character}(object, pattern, replacement, ...)
}
\arguments{
\item{object}{generic object (e.g. model, code_record(s), etc.)}

\item{pattern}{pattern to be replaced}

\item{replacement}{replacement string}

\item{...}{extra arguments}
}
\value{
the same object with all occurrences replaced
}
\description{
Replace all occurrences in object.
}
