% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{replaceAll}
\alias{replaceAll}
\title{Replace all occurrences in object.}
\usage{
replaceAll(object, pattern, replacement, ...)
}
\arguments{
\item{object}{generic object (e.g. model, code_record(s), etc.)}

\item{pattern}{pattern to be replaced}

\item{replacement}{replacement string}

\item{...}{extra arguments}
}
\value{
the loaded S4 object
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`replaceAll()` is deprecated in favor of `replace_all()`.
}
