% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{replaceEquation}
\alias{replaceEquation}
\title{Replace equation in code records.}
\usage{
replaceEquation(object, lhs, rhs)
}
\arguments{
\item{object}{object containing code records}

\item{lhs}{left-hand-side variable}

\item{rhs}{right-hand-side expression}
}
\value{
updated code records
}
\description{
Replace equation in code records.
}
