% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_settings.R
\docType{class}
\name{replication_settings-class}
\alias{replication_settings-class}
\title{Replication settings class.}
\description{
Replication settings class.
}
\section{Slots}{

\describe{
\item{\code{wishart}}{logical, sample OMEGAs and SIGMAs from inverse Chi-Square or Wishart distributions}

\item{\code{nsub}}{number of subjects}

\item{\code{nobs}}{number of observations}
}}

