% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_settings.R
\name{ReplicationSettings}
\alias{ReplicationSettings}
\title{Create replication settings.}
\usage{
ReplicationSettings(wishart = FALSE, nsub = NA, nobs = NA)
}
\arguments{
\item{wishart}{logical, sample OMEGAs and SIGMAs from inverse Chi-Square or Wishart distributions}

\item{nsub}{number of subjects}

\item{nobs}{number of observations}
}
\value{
replication settings
}
\description{
Create replication settings.
}
