% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode_conversion.R
\name{rxode_code}
\alias{rxode_code}
\title{Get code for rxode2}
\usage{
rxode_code(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
corresponding model code for rxode2
}
\description{
Get code for rxode2
}
