% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode_conversion.R
\name{rxode_matrix}
\alias{rxode_matrix}
\title{Get the OMEGA/SIGMA matrix for rxode2.}
\usage{
rxode_matrix(model, type = "omega")
}
\arguments{
\item{model}{Campsis model or Campsis parameters}

\item{type}{either omega or sigma}
}
\value{
omega/sigma named matrix
}
\description{
Get the OMEGA/SIGMA matrix for rxode2.
}
