% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode_conversion.R
\name{rxode_params}
\alias{rxode_params}
\title{Get the parameters vector for rxode2.}
\usage{
rxode_params(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
named vector with THETA values
}
\description{
Get the parameters vector for rxode2.
}
