% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode_conversion.R
\name{rxodeParams}
\alias{rxodeParams}
\title{Get the parameters vector for RxODE.}
\usage{
rxodeParams(pmxmod)
}
\arguments{
\item{pmxmod}{PMX model}
}
\value{
named vector with THETA values
}
\description{
Get the parameters vector for RxODE.
}
