% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{sample_from_inverse_chi_squared_core}
\alias{sample_from_inverse_chi_squared_core}
\title{Sample from scaled inverse chi-squared distribution (core method).}
\usage{
sample_from_inverse_chi_squared_core(n, df, scale, variable)
}
\arguments{
\item{n}{numbers of rows to sample}

\item{df}{degree of freedom}

\item{scale}{scale parameter}

\item{variable}{variable name}
}
\value{
a data frame with the unique sampled parameter
}
\description{
Sample from scaled inverse chi-squared distribution (core method).
}
\keyword{internal}
