% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{sample_from_inverse_chi_squared_or_wishart}
\alias{sample_from_inverse_chi_squared_or_wishart}
\title{Sample parameters from inverse scaled chi-squared or wishart distribution(s).}
\usage{
sample_from_inverse_chi_squared_or_wishart(parameters, n, settings)
}
\arguments{
\item{parameters}{subset of Campsis parameters (OMEGAs or SIGMAs)}

\item{n}{number of rows to sample}

\item{settings}{replication settings}
}
\value{
a data frame with the sampled parameters
}
\description{
Sample parameters from inverse scaled chi-squared or wishart distribution(s).
}
\keyword{internal}
