% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{sample_from_inverse_wishart_core}
\alias{sample_from_inverse_wishart_core}
\title{Sample from scaled inverse Wishart distribution (core method).}
\usage{
sample_from_inverse_wishart_core(n, df, mat, allColnames, wishartCorrection)
}
\arguments{
\item{n}{numbers of rows to sample}

\item{df}{degree of freedom}

\item{mat}{scaling matrix}

\item{allColnames}{all column names as they are going to appear when as.vector is called}

\item{wishartCorrection}{apply correction to the degrees of freedom, logical}
}
\value{
a data frame with the sampled parameters
}
\description{
Sample from scaled inverse Wishart distribution (core method).
}
\keyword{internal}
