% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{sample_from_multivariate_normal_distribution_core}
\alias{sample_from_multivariate_normal_distribution_core}
\title{Sample from multivariate normal distribution (core method).}
\usage{
sample_from_multivariate_normal_distribution_core(n, mean, varcov)
}
\arguments{
\item{n}{numbers of rows to sample}

\item{mean}{mean values to give to the multivariate normal distribution}

\item{varcov}{variance-covariance matrix}
}
\value{
a data frame with the sampled parameters
}
\description{
Sample from multivariate normal distribution (core method).
}
\keyword{internal}
