% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{sampleFromInverseChiSquaredOrWishart}
\alias{sampleFromInverseChiSquaredOrWishart}
\title{Sample parameters from inverse scaled chi-squared or wishart distribution(s).}
\usage{
sampleFromInverseChiSquaredOrWishart(parameters, n, df)
}
\arguments{
\item{parameters}{subset of Campsis parameters (OMEGAs or SIGMAs)}

\item{n}{number of rows to sample}

\item{df}{degree of freedom for the scaled inverse chi-squared or wishart distribution}
}
\value{
a data frame with the sampled parameters
}
\description{
Sample parameters from inverse scaled chi-squared or wishart distribution(s).
}
