% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{sampleFromInverseWishartCore}
\alias{sampleFromInverseWishartCore}
\title{Sample from scaled inverse Wishart distribution (core method).}
\usage{
sampleFromInverseWishartCore(n, df, mat, allColnames)
}
\arguments{
\item{n}{numbers of rows to sample}

\item{df}{degree of freedom}

\item{mat}{scaling matrix}

\item{allColnames}{all column names as they are going to appear when as.vector is called}
}
\value{
a data frame with the sampled parameters
}
\description{
Sample from scaled inverse Wishart distribution (core method).
}
\keyword{internal}
