% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{sampleFromMultivariateNormalDistribution}
\alias{sampleFromMultivariateNormalDistribution}
\title{Sample from a multivariate normal distribution.}
\usage{
sampleFromMultivariateNormalDistribution(parameters, n, settings)
}
\arguments{
\item{parameters}{model parameters (all), already standardised}

\item{n}{number of rows to sample}

\item{settings}{replication settings}
}
\value{
a data frame with the sampled parameters
}
\description{
Sample from a multivariate normal distribution.
}
\keyword{internal}
