% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{sampleGeneric}
\alias{sampleGeneric}
\title{Generic function for parameter sampling according to the minimum and maximum values.
This function will sample parameters a first time and check if some parameters are out of range.
Based on the success rate, it will sample more parameters to reach the desired number of rows.}
\usage{
sampleGeneric(fun, args, n, minMax, msg)
}
\arguments{
\item{n}{number of rows to sample}

\item{msg}{message template}

\item{parameters}{Campsis parameters present in the variance-covariance matrix}

\item{varcov}{variance-covariance matrix}
}
\value{
description
}
\description{
Generic function for parameter sampling according to the minimum and maximum values.
This function will sample parameters a first time and check if some parameters are out of range.
Based on the success rate, it will sample more parameters to reach the desired number of rows.
}
