% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{sampleMore}
\alias{sampleMore}
\title{Sample more parameters.}
\usage{
sampleMore(n, mean, varcov, shift)
}
\arguments{
\item{n}{numbers of rows to sample}

\item{mean}{mean values to give to the multivariate normal distribution}

\item{varcov}{variance-covariance matrix}

\item{shift}{shift to apply to the replicate ID column}
}
\description{
Sample more parameters.
}
