% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicated_campsis_model.R
\name{sampleOmegasSigmas}
\alias{sampleOmegasSigmas}
\title{Sample the OMEGAs and SIGMAs from scaled inverse chi-square or wishart distributions.}
\usage{
sampleOmegasSigmas(model, type = "omega", n, df)
}
\arguments{
\item{model}{Campsis model}

\item{type}{type of parameter to sample (omega or sigma)}

\item{n}{number of rows to sample}

\item{df}{degree of freedom for the scaled inverse chi-square or wishart distribution}
}
\value{
a data frame with the sampled parameters
}
\description{
Sample the OMEGAs and SIGMAs from scaled inverse chi-square or wishart distributions.
}
