% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/compartment_properties.R,
%   R/parameters.R
\name{select}
\alias{select}
\alias{select,data.frame-method}
\alias{select,compartment_properties-method}
\alias{select,parameters-method}
\title{Select something from given object.}
\usage{
select(object, ...)

\S4method{select}{data.frame}(object, ...)

\S4method{select}{compartment_properties}(object, ...)

\S4method{select}{parameters}(object, ...)
}
\arguments{
\item{object}{generic object}

\item{...}{arguments to select}
}
\value{
filtered object
}
\description{
Select something from given object.
}
