% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_block.R
\name{shift_omega_indexes}
\alias{shift_omega_indexes}
\alias{shift_omega_indexes,omega_block-method}
\title{Shift OMEGA indexes.}
\usage{
shift_omega_indexes(object)

\S4method{shift_omega_indexes}{omega_block}(object)
}
\arguments{
\item{object}{omega block}
}
\value{
same block but shifted
}
\description{
Shift OMEGA indexes.
}
