% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_block.R
\name{shiftOmegaIndexes}
\alias{shiftOmegaIndexes}
\alias{shiftOmegaIndexes,omega_block-method}
\title{Shift OMEGA indexes.}
\usage{
shiftOmegaIndexes(object)

\S4method{shiftOmegaIndexes}{omega_block}(object)
}
\arguments{
\item{object}{omega block}
}
\value{
same block but shifted
}
\description{
Shift OMEGA indexes.
}
