% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\docType{class}
\name{single_array_parameter-class}
\alias{single_array_parameter-class}
\title{Single-array parameter class. This parameter has a single index value.}
\description{
Single-array parameter class. This parameter has a single index value.
}
