% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R, R/compartment_properties.R,
%   R/compartments.R, R/parameters.R, R/code_records.R, R/campsis_model.R
\name{sort}
\alias{sort}
\alias{sort,compartment_properties-method}
\alias{sort,compartments-method}
\alias{sort,parameters-method}
\alias{sort,code_records-method}
\alias{sort,campsis_model-method}
\title{Sort the specified list.}
\usage{
sort(x, decreasing = FALSE, ...)

\S4method{sort}{compartment_properties}(x, decreasing = FALSE, ...)

\S4method{sort}{compartments}(x, decreasing = FALSE, ...)

\S4method{sort}{parameters}(x, decreasing = FALSE, ...)

\S4method{sort}{code_records}(x, decreasing = FALSE, ...)

\S4method{sort}{campsis_model}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{list object}

\item{decreasing}{increasing or decreasing order}

\item{...}{extra arguments}
}
\value{
same list but ordered
}
\description{
Sort the specified list.
}
