% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_record.R
\docType{class}
\name{statement_record-class}
\alias{statement_record-class}
\title{Statement record class.}
\description{
Statement record class.
}
\section{Slots}{

\describe{
\item{\code{statements}}{model statements}
}}

