% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{Theta}
\alias{Theta}
\title{Create a THETA parameter.}
\usage{
Theta(
  name = NA,
  index = NA,
  value = NA,
  min = NA,
  max = NA,
  fix = FALSE,
  label = NA,
  unit = NA,
  comment = NA
)
}
\arguments{
\item{name}{parameter name, e.g. CL (prefix THETA will be added automatically)}

\item{index}{parameter index}

\item{value}{parameter value}

\item{min}{minimum value for this parameter when parameter uncertainty is enabled}

\item{max}{maximum value for this parameter when parameter uncertainty is enabled}

\item{fix}{parameter was fixed in estimation, logical value}

\item{label}{parameter label, optional}

\item{unit}{parameter unit, optional}

\item{comment}{any comment, optional}
}
\value{
a THETA parameter
}
\description{
Create a THETA parameter.
}
