% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/model_unknown_statement.R,
%   R/model_comment.R, R/model_line_break.R, R/model_equation.R, R/model_ode.R,
%   R/model_if_statement.R, R/compartment.R, R/compartment_property.R,
%   R/compartment_initial_condition.R, R/mrgsolve_model.R
\name{to_string}
\alias{to_string}
\alias{to_string,unknown_statement-method}
\alias{to_string,comment-method}
\alias{to_string,line_break-method}
\alias{to_string,equation-method}
\alias{to_string,ode-method}
\alias{to_string,if_statement-method}
\alias{to_string,compartment-method}
\alias{to_string,compartment_property-method}
\alias{to_string,compartment_initial_condition-method}
\alias{to_string,mrgsolve_model-method}
\title{to_string generic method.}
\usage{
to_string(object, ...)

\S4method{to_string}{unknown_statement}(object, ...)

\S4method{to_string}{comment}(object, ...)

\S4method{to_string}{line_break}(object, ...)

\S4method{to_string}{equation}(object, ...)

\S4method{to_string}{ode}(object, ...)

\S4method{to_string}{if_statement}(object, ...)

\S4method{to_string}{compartment}(object, ...)

\S4method{to_string}{compartment_property}(object, ...)

\S4method{to_string}{compartment_initial_condition}(object, ...)

\S4method{to_string}{mrgsolve_model}(object, ...)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments needed for to_string conversion}
}
\value{
character value/vector
}
\description{
to_string generic method.
}
