% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trim}
\alias{trim}
\title{Trim character vector. Remove all leading and trailing spaces.}
\usage{
trim(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector without leading and trailing spaces
}
\description{
Trim character vector. Remove all leading and trailing spaces.
}
