% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_element_position.R
\name{UndefinedPosition}
\alias{UndefinedPosition}
\title{Undefined position.}
\usage{
UndefinedPosition()
}
\value{
an undefined position object
}
\description{
Undefined position.
}
\keyword{internal}
