% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_unknown_statement.R
\docType{class}
\name{unknown_statement-class}
\alias{unknown_statement-class}
\title{Unknown statement class. Any statement not recognized by campsismod.}
\description{
Unknown statement class. Any statement not recognized by campsismod.
}
