% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_unknown_statement.R
\name{UnknownStatement}
\alias{UnknownStatement}
\title{Create a new ordinary differential equation (ODE).}
\usage{
UnknownStatement(line, comment = as.character(NA))
}
\arguments{
\item{line}{line which was not recognised}

\item{comment}{comment if any, single character string}
}
\value{
an unknown statement
}
\description{
Create a new ordinary differential equation (ODE).
}
