% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_statement.R
\name{UnsupportedDestException}
\alias{UnsupportedDestException}
\title{Unknown destination engine exception.}
\usage{
UnsupportedDestException()
}
\value{
no return value
}
\description{
Unknown destination engine exception.
}
\keyword{internal}
