% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campsis_model.R
\name{update_compartments}
\alias{update_compartments}
\title{Update compartments list from the persisted records.
Exported especially for package \code{campsistrans}.
However, this method should not be called.}
\usage{
update_compartments(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
an updated Campsis model, with an updated compartments list
}
\description{
Update compartments list from the persisted records.
Exported especially for package \code{campsistrans}.
However, this method should not be called.
}
