% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicated_campsis_model.R
\name{update_omegas}
\alias{update_omegas}
\title{Update OMEGAs that are same. Same OMEGAs are written as follows:
OMEGA1 same is FALSE (first one, estimated)
OMEGA2 same is TRUE
OMEGA3 same is TRUE, etc.
OMEGA2 and OMEGA3 will take the same value as OMEGA1.}
\usage{
update_omegas(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
updated Campsis model
}
\description{
Update OMEGAs that are same. Same OMEGAs are written as follows:
OMEGA1 same is FALSE (first one, estimated)
OMEGA2 same is TRUE
OMEGA3 same is TRUE, etc.
OMEGA2 and OMEGA3 will take the same value as OMEGA1.
}
\keyword{internal}
