% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{updateCompartments}
\alias{updateCompartments}
\title{Update compartments list from the persisted records.
Exported especially for package \code{campsistrans}.
However, this method should not be called.}
\usage{
updateCompartments(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
an updated Campsis model, with an updated compartments list
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`updateCompartments()` is deprecated in favor of `update_compartments()`.
}
