% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicated_campsis_model.R
\name{updateParameters}
\alias{updateParameters}
\title{Update model parameters based on the parameters issued from the model replication.}
\usage{
updateParameters(model, row)
}
\arguments{
\item{model}{Campsis model}

\item{row}{a data frame row containing the new parameter values}
}
\value{
updated Campsis model
}
\description{
Update model parameters based on the parameters issued from the model replication.
}
\keyword{internal}
