% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{utilsGlobalFunctions}
\alias{utilsGlobalFunctions}
\title{A empty method only used to declare some of the global function definitions
contained in the package 'utils'.}
\usage{
utilsGlobalFunctions()
}
\description{
A empty method only used to declare some of the global function definitions
contained in the package 'utils'.
}
\keyword{internal}
