% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{variable_pattern_no_start_str}
\alias{variable_pattern_no_start_str}
\title{Return the variable pattern (string form), without the first character.}
\usage{
variable_pattern_no_start_str()
}
\value{
pattern (regular expression)
}
\description{
Return the variable pattern (string form), without the first character.
}
\keyword{internal}
