% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{variable_pattern_str}
\alias{variable_pattern_str}
\title{Return the variable pattern (string form).}
\usage{
variable_pattern_str()
}
\value{
pattern (regular expression)
}
\description{
Return the variable pattern (string form).
}
\keyword{internal}
