% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{VariablePattern}
\alias{VariablePattern}
\title{Create a variable pattern.}
\usage{
VariablePattern(x)
}
\arguments{
\item{x}{variable name}
}
\value{
a pattern
}
\description{
Create a variable pattern.
}
