% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{variablePatternNoStartStr}
\alias{variablePatternNoStartStr}
\title{Return the variable pattern (string form), without the first character.}
\usage{
variablePatternNoStartStr()
}
\value{
pattern (regular expression)
}
\description{
Return the variable pattern (string form), without the first character.
}
\keyword{internal}
