% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{variablePatternStr}
\alias{variablePatternStr}
\title{Return the variable pattern (string form).}
\usage{
variablePatternStr()
}
\value{
pattern (regular expression)
}
\description{
Return the variable pattern (string form).
}
\keyword{internal}
