% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{write_parameters}
\alias{write_parameters}
\title{Write subset of parameters (theta, omega or sigma).}
\usage{
write_parameters(object, file, ...)
}
\arguments{
\item{object}{subset of parameters}

\item{file}{filename}

\item{...}{extra arguments, like \code{defaultDf} for empty parameters list}
}
\value{
TRUE if success
}
\description{
Write subset of parameters (theta, omega or sigma).
}
