% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_record.R
\name{write_record_delimiter}
\alias{write_record_delimiter}
\title{Write record delimiter line.}
\usage{
write_record_delimiter(object)
}
\arguments{
\item{object}{code record}
}
\value{
a record delimiter line
}
\description{
Write record delimiter line.
}
\keyword{internal}
