% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{write_varcov}
\alias{write_varcov}
\title{Write variance-covariance matrix.}
\usage{
write_varcov(object, file)
}
\arguments{
\item{object}{matrix}

\item{file}{filename}
}
\value{
TRUE if success
}
\description{
Write variance-covariance matrix.
}
