% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_record.R
\name{writeRecordDelimiter}
\alias{writeRecordDelimiter}
\title{Write record delimiter line.}
\usage{
writeRecordDelimiter(object)
}
\arguments{
\item{object}{code record}
}
\value{
a record delimiter line
}
\description{
Write record delimiter line.
}
\keyword{internal}
