% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{writeVarcov}
\alias{writeVarcov}
\title{Write variance-covariance matrix.}
\usage{
writeVarcov(object, file)
}
\arguments{
\item{object}{matrix}

\item{file}{filename}
}
\value{
TRUE if success
}
\description{
Write variance-covariance matrix.
}
