% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{addCharacteristic}
\alias{addCharacteristic}
\title{Add characteristic to compartments object.}
\usage{
addCharacteristic(line, emptyCharacteristic, compartments)
}
\arguments{
\item{line}{line record}

\item{emptyCharacteristic}{empty characteristic, to be completed}

\item{compartments}{compartments object}
}
\value{
updated compartments object
}
\description{
Add characteristic to compartments object.
}
