% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/code_record.R, R/code_records.R,
%   R/campsis_model.R
\name{addEquation}
\alias{addEquation}
\alias{addEquation,code_record,character,character-method}
\alias{addEquation,code_records,character,character-method}
\alias{addEquation,campsis_model,character,character-method}
\title{Add equation.}
\usage{
addEquation(object, lhs, rhs, ...)

\S4method{addEquation}{code_record,character,character}(object, lhs, rhs, before = NULL, after = NULL)

\S4method{addEquation}{code_records,character,character}(object, lhs, rhs, before = NULL, after = NULL)

\S4method{addEquation}{campsis_model,character,character}(object, lhs, rhs, before = NULL, after = NULL)
}
\arguments{
\item{object}{object containing code record(s)}

\item{lhs}{left-hand-side variable}

\item{rhs}{right-hand-side expression}

\item{...}{extra arguments like before or after}

\item{before}{index or variable, may be used to insert an equation at a specific position, before this index (in record) or variable}

\item{after}{index or variable, may be used to insert an equation at a specific position, after this index (in record) or variable}
}
\value{
updated object
}
\description{
Add equation.
}
