% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_settings.R
\docType{class}
\name{auto_replication_settings-class}
\alias{auto_replication_settings-class}
\title{Auto replication settings class.}
\description{
Auto replication settings class.
}
\section{Slots}{

\describe{
\item{\code{wishart}}{logical, sample OMEGAs and SIGMAs from scaled inverse chi-squared or Wishart distributions}

\item{\code{nsub}}{number of subjects}

\item{\code{nobs}}{number of observations}

\item{\code{quiet}}{logical, suppress info messages}
}}

