% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_settings.R
\docType{class}
\name{auto_replication_settings-class}
\alias{auto_replication_settings-class}
\title{Auto replication settings class.}
\description{
Auto replication settings class.
}
\section{Slots}{

\describe{
\item{\code{wishart}}{logical, sample OMEGAs and SIGMAs from scaled inverse chi-squared or Wishart distributions}

\item{\code{odf}}{the degrees of freedom for the scaled inverse chi-squared/Wishart distribution with regards to the OMEGAs}

\item{\code{sdf}}{the degrees of freedom for the scaled inverse chi-squared/Wishart distribution with regards to the SIGMAs}

\item{\code{quiet}}{logical, suppress info messages}

\item{\code{max_iterations}}{number of iterations maximum to sample the parameters}

\item{\code{max_chunk_size}}{maximum number of rows to sample at once}

\item{\code{min_max}}{logical, check for min/max values when sampling the parameters}

\item{\code{positive_definite}}{logical, check for positive definiteness when sampling the OMEGA/SIGMA parameters}
}}

