% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_settings.R
\name{AutoReplicationSettings}
\alias{AutoReplicationSettings}
\title{Create auto replication settings.
By default, all model parameters are sampled from a multivariate normal 
distribution, whose characteristics are specified by the variance-covariance matrix.
OMEGAs and SIGMAs can be sampled from scaled inverse chi-squared or Wishart distributions
by setting the 'wishart' argument to 'TRUE'. In that case, THETAs are still sampled
from a multivariate normal distribution; while OMEGAS and SIGMAs are sampled from
scaled inverse chi-squared (univariate OMEGA/SIGMA distribution) and Wishart (block of OMEGA)
distribution, respectively. When 'wishart' is set to 'TRUE', the 'nsub' (number of 
subjects in modelling) and 'nobs' (total number of observations in modelling) arguments
must be specified.}
\usage{
AutoReplicationSettings(wishart = FALSE, nsub = NA, nobs = NA, quiet = NA)
}
\arguments{
\item{wishart}{logical, sample OMEGAs and SIGMAs from scaled inverse chi-squared (univariate OMEGA distribution)
or Wishart distribution (block of OMEGAs)}

\item{nsub}{number of subjects in modelling, used as the degree of freedom for the scaled inverse chi-squared/Wishart distribution for OMEGAs}

\item{nobs}{number of observations in modelling, used as the degree of freedom for the scaled inverse chi-squared/Wishart distribution for SIGMAs}

\item{quiet}{logical, suppress info messages, default is NA. By default, messages will be printed out when the success rate of sampling the parameters is below 95\%.}
}
\value{
replication settings
}
\description{
Create auto replication settings.
By default, all model parameters are sampled from a multivariate normal 
distribution, whose characteristics are specified by the variance-covariance matrix.
OMEGAs and SIGMAs can be sampled from scaled inverse chi-squared or Wishart distributions
by setting the 'wishart' argument to 'TRUE'. In that case, THETAs are still sampled
from a multivariate normal distribution; while OMEGAS and SIGMAs are sampled from
scaled inverse chi-squared (univariate OMEGA/SIGMA distribution) and Wishart (block of OMEGA)
distribution, respectively. When 'wishart' is set to 'TRUE', the 'nsub' (number of 
subjects in modelling) and 'nobs' (total number of observations in modelling) arguments
must be specified.
}
