% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_bioavailability.R
\name{Bioavailability}
\alias{Bioavailability}
\title{Create a bioavailability for the specified compartment.}
\usage{
Bioavailability(compartment, rhs = "")
}
\arguments{
\item{compartment}{compartment index}

\item{rhs}{right-hand side part of the equation}
}
\value{
a bioavailability property
}
\description{
Create a bioavailability for the specified compartment.
}
