% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comment.R
\docType{class}
\name{comment-class}
\alias{comment-class}
\title{Equation class. Any statement in the form A = B.}
\description{
Equation class. Any statement in the form A = B.
}
\section{Slots}{

\describe{
\item{\code{comment}}{a comment associated to this model statement}
}}

