% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R, R/code_record.R,
%   R/campsis_model.R
\name{contains}
\alias{contains}
\alias{contains,pmx_list,pmx_element-method}
\alias{contains,statements_record,model_statement-method}
\alias{contains,campsis_model,pmx_element-method}
\title{Check if an element exists in list.}
\usage{
contains(object, x)

\S4method{contains}{pmx_list,pmx_element}(object, x)

\S4method{contains}{statements_record,model_statement}(object, x)

\S4method{contains}{campsis_model,pmx_element}(object, x)
}
\arguments{
\item{object}{list object}

\item{x}{element to check if exists}
}
\value{
logical value, TRUE or FALSE
}
\description{
Check if an element exists in list.
}
