% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/parameters.R, R/pmx_model.R
\name{disable}
\alias{disable}
\alias{disable,parameters,character-method}
\alias{disable,pmx_model,character-method}
\title{Disable.}
\usage{
disable(object, x, ...)

\S4method{disable}{parameters,character}(object, x, ...)

\S4method{disable}{pmx_model,character}(object, x, ...)
}
\arguments{
\item{object}{generic object}

\item{x}{what needs to be disabled}

\item{...}{extra arguments needed for disabling}
}
\value{
standardised object
}
\description{
Disable.
}
