% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/campsis_model.R, R/rxode_model.R,
%   R/mrgsolve_model.R
\name{export}
\alias{export}
\alias{export,campsis_model,character-method}
\alias{export,campsis_model,rxode_type-method}
\alias{export,campsis_model,mrgsolve_type-method}
\title{Export function.}
\usage{
export(object, dest, ...)

\S4method{export}{campsis_model,character}(object, dest, ...)

\S4method{export}{campsis_model,rxode_type}(object, dest, ...)

\S4method{export}{campsis_model,mrgsolve_type}(object, dest, outvars = NULL, extra_params = character(0))
}
\arguments{
\item{object}{generic object}

\item{dest}{destination}

\item{...}{optional arguments}

\item{outvars}{additional variables to capture}

\item{extra_params}{extra parameter names to be added. By default, they will be assigned a zero value.}
}
\value{
specific object depending on given destination
}
\description{
Export function.
}
