% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extractLhs}
\alias{extractLhs}
\title{Extract left-hand-side expression.}
\usage{
extractLhs(x)
}
\arguments{
\item{x}{character value}
}
\value{
left-hand-side expression, not trimmed
}
\description{
Extract left-hand-side expression.
}
