% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extractRhs}
\alias{extractRhs}
\title{Extract right-hand side expression.}
\usage{
extractRhs(x)
}
\arguments{
\item{x}{character value}
}
\value{
right-hand side expressionn
}
\description{
Extract right-hand side expression.
}
