% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R, R/compartments.R,
%   R/code_record.R, R/code_records.R, R/campsis_model.R
\name{find}
\alias{find}
\alias{find,pmx_list,pmx_element-method}
\alias{find,compartments,compartment_property-method}
\alias{find,statements_record,model_statement-method}
\alias{find,code_records,model_statement-method}
\alias{find,campsis_model,compartment-method}
\alias{find,campsis_model,compartment_property-method}
\alias{find,campsis_model,parameter-method}
\alias{find,campsis_model,code_record-method}
\alias{find,campsis_model,model_statement-method}
\title{Find an element in list.}
\usage{
find(object, x)

\S4method{find}{pmx_list,pmx_element}(object, x)

\S4method{find}{compartments,compartment_property}(object, x)

\S4method{find}{statements_record,model_statement}(object, x)

\S4method{find}{code_records,model_statement}(object, x)

\S4method{find}{campsis_model,compartment}(object, x)

\S4method{find}{campsis_model,compartment_property}(object, x)

\S4method{find}{campsis_model,parameter}(object, x)

\S4method{find}{campsis_model,code_record}(object, x)

\S4method{find}{campsis_model,model_statement}(object, x)
}
\arguments{
\item{object}{list object}

\item{x}{element to find, only key slots need to be filled in}
}
\value{
the element from the list that has same name as x, or NULL if no element was found
}
\description{
Find an element in list.
}
