% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{fixOmega}
\alias{fixOmega}
\title{Fix omega matrix for SAME OMEGA parameters that have NA values due to imperfections in Pharmpy import.}
\usage{
fixOmega(object)
}
\arguments{
\item{object}{generic object}
}
\value{
the parameter that matches
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`fixOmega()` is deprecated in favor of `fix_omega()`.
}
