% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{fixOmega}
\alias{fixOmega}
\alias{fixOmega,parameters-method}
\title{Fix omega matrix for SAME OMEGA parameters that have NA values due to imperfections in Pharmpy import.}
\usage{
fixOmega(object)

\S4method{fixOmega}{parameters}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
the parameter that matches
}
\description{
Fix omega matrix for SAME OMEGA parameters that have NA values due to imperfections in Pharmpy import.
}
