% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_element_list.R,
%   R/model_unknown_statement.R, R/model_comment.R, R/model_line_break.R,
%   R/model_equation.R, R/model_ode.R, R/model_if_statement.R, R/compartment.R,
%   R/compartment_bioavailability.R, R/compartment_lag_time.R,
%   R/compartment_infusion_duration.R, R/compartment_infusion_rate.R,
%   R/compartment_initial_condition.R, R/parameter.R, R/code_record.R,
%   R/omega_block.R
\name{get_name}
\alias{get_name}
\alias{get_name,ANY-method}
\alias{get_name,unknown_statement-method}
\alias{get_name,comment-method}
\alias{get_name,line_break-method}
\alias{get_name,equation-method}
\alias{get_name,ode-method}
\alias{get_name,if_statement-method}
\alias{get_name,compartment-method}
\alias{get_name,compartment_bioavailability-method}
\alias{get_name,compartment_lag_time-method}
\alias{get_name,compartment_infusion_duration-method}
\alias{get_name,compartment_infusion_rate-method}
\alias{get_name,compartment_initial_condition-method}
\alias{get_name,theta-method}
\alias{get_name,omega-method}
\alias{get_name,sigma-method}
\alias{get_name,main_record-method}
\alias{get_name,ode_record-method}
\alias{get_name,f_record-method}
\alias{get_name,lag_record-method}
\alias{get_name,duration_record-method}
\alias{get_name,rate_record-method}
\alias{get_name,init_record-method}
\alias{get_name,error_record-method}
\alias{get_name,omega_block-method}
\title{Get element name.}
\usage{
get_name(x)

\S4method{get_name}{ANY}(x)

\S4method{get_name}{unknown_statement}(x)

\S4method{get_name}{comment}(x)

\S4method{get_name}{line_break}(x)

\S4method{get_name}{equation}(x)

\S4method{get_name}{ode}(x)

\S4method{get_name}{if_statement}(x)

\S4method{get_name}{compartment}(x)

\S4method{get_name}{compartment_bioavailability}(x)

\S4method{get_name}{compartment_lag_time}(x)

\S4method{get_name}{compartment_infusion_duration}(x)

\S4method{get_name}{compartment_infusion_rate}(x)

\S4method{get_name}{compartment_initial_condition}(x)

\S4method{get_name}{theta}(x)

\S4method{get_name}{omega}(x)

\S4method{get_name}{sigma}(x)

\S4method{get_name}{main_record}(x)

\S4method{get_name}{ode_record}(x)

\S4method{get_name}{f_record}(x)

\S4method{get_name}{lag_record}(x)

\S4method{get_name}{duration_record}(x)

\S4method{get_name}{rate_record}(x)

\S4method{get_name}{init_record}(x)

\S4method{get_name}{error_record}(x)

\S4method{get_name}{omega_block}(x)
}
\arguments{
\item{x}{element to know the name}
}
\value{
the name of this element
}
\description{
Get element name.
}
