% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{getByIndex}
\alias{getByIndex}
\title{Get element by index.}
\usage{
getByIndex(object, x)
}
\arguments{
\item{object}{list object}

\item{x}{element index}
}
\value{
index of this element
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`getByIndex()` is deprecated in favor of `get_by_index()`.
}
\keyword{internal}
