% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{getByIndex}
\alias{getByIndex}
\title{Get parameter by index (single index).}
\usage{
getByIndex(object, parameter)
}
\arguments{
\item{object}{list of parameters}

\item{parameter}{to search for}
}
\value{
parameter that matches
}
\description{
Get parameter by index (single index).
}
