% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\name{getByIndex}
\alias{getByIndex}
\title{Get element by index.}
\usage{
getByIndex(object, x)
}
\arguments{
\item{object}{list object}

\item{x}{something to search for, either indexed element or index itself}
}
\value{
element from the list whose index matches with provided index
}
\description{
Get element by index.
}
