% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{getCompartmentIndex}
\alias{getCompartmentIndex}
\title{Get the compartment index for the specified compartment name.}
\usage{
getCompartmentIndex(object, name)
}
\arguments{
\item{object}{generic object that contains compartments information}

\item{name}{compartment name}
}
\value{
the corresponding compartment index
}
\description{
Get the compartment index for the specified compartment name.
}
