% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/compartments.R, R/campsis_model.R
\name{getCompartmentIndex}
\alias{getCompartmentIndex}
\alias{getCompartmentIndex,compartments,character-method}
\alias{getCompartmentIndex,campsis_model,character-method}
\title{Get the compartment index for the specified compartment name.}
\usage{
getCompartmentIndex(object, name)

\S4method{getCompartmentIndex}{compartments,character}(object, name)

\S4method{getCompartmentIndex}{campsis_model,character}(object, name)
}
\arguments{
\item{object}{generic object that contains compartments information}

\item{name}{compartment name}
}
\value{
the corresponding compartment index
}
\description{
Get the compartment index for the specified compartment name.
}
