% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{getCompartments}
\alias{getCompartments}
\title{Detect all compartments names from the code records.
Only for model instantiation. Not exported.}
\usage{
getCompartments(records)
}
\arguments{
\item{records}{code records}
}
\value{
a list of compartments
}
\description{
Detect all compartments names from the code records.
Only for model instantiation. Not exported.
}
