% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{getCompartments}
\alias{getCompartments}
\title{Detect all compartments from model code records.
Only for model instantiation. Not exported.}
\usage{
getCompartments(records)
}
\arguments{
\item{records}{code records}
}
\value{
a list of 2 args: first argument is the ODE record without characteristics,
second argument is the compartments object
}
\description{
Detect all compartments from model code records.
Only for model instantiation. Not exported.
}
