% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{getCompartments}
\alias{getCompartments}
\title{Get all compartments from model.}
\usage{
getCompartments(object)
}
\arguments{
\item{object}{generic object}
}
\value{
a list of compartments
}
\description{
Get all compartments from model.
}
