% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/code_record.R, R/code_records.R,
%   R/campsis_model.R
\name{getEquation}
\alias{getEquation}
\alias{getEquation,code_record,character-method}
\alias{getEquation,code_records,character-method}
\alias{getEquation,campsis_model,character-method}
\title{Get equation from code record(s).}
\usage{
getEquation(object, lhs)

\S4method{getEquation}{code_record,character}(object, lhs)

\S4method{getEquation}{code_records,character}(object, lhs)

\S4method{getEquation}{campsis_model,character}(object, lhs)
}
\arguments{
\item{object}{object containing code record(s)}

\item{lhs}{left-hand-side variable}
}
\value{
equation, character or NULL if not found
}
\description{
Get equation from code record(s).
}
