% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{getName}
\alias{getName}
\title{Get element names from list.}
\usage{
getName(x)
}
\arguments{
\item{x}{element to know the name}
}
\value{
character vector with all the element names of this list
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`getName()` is deprecated in favor of `get_name()`.
}
\keyword{internal}
