% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{getNameInModel}
\alias{getNameInModel}
\title{Get name of parameter in the PMX model.}
\usage{
getNameInModel(x)
}
\arguments{
\item{x}{element to know the name}
}
\value{
the name of this element
}
\description{
Get name of parameter in the PMX model.
}
