% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{getNameInModel}
\alias{getNameInModel}
\alias{getNameInModel,theta-method}
\alias{getNameInModel,omega-method}
\alias{getNameInModel,sigma-method}
\title{Get the name of the given parameter in the Campsis model.}
\usage{
getNameInModel(x)

\S4method{getNameInModel}{theta}(x)

\S4method{getNameInModel}{omega}(x)

\S4method{getNameInModel}{sigma}(x)
}
\arguments{
\item{x}{element to know the name}
}
\value{
the name of this parameter
}
\description{
Get the name of the given parameter in the Campsis model.
}
