% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\name{getNames}
\alias{getNames}
\alias{getNames,pmx_list-method}
\title{Get element names from list.}
\usage{
getNames(object)

\S4method{getNames}{pmx_list}(object)
}
\arguments{
\item{object}{list object}
}
\value{
character vector
}
\description{
Get element names from list.
}
