% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_property.R,
%   R/compartment_bioavailability.R, R/compartment_lag_time.R,
%   R/compartment_infusion_duration.R, R/compartment_infusion_rate.R,
%   R/compartment_initial_condition.R
\name{getPrefix}
\alias{getPrefix}
\alias{getPrefix,compartment_bioavailability-method}
\alias{getPrefix,compartment_lag_time-method}
\alias{getPrefix,compartment_infusion_duration-method}
\alias{getPrefix,compartment_infusion_rate-method}
\alias{getPrefix,compartment_initial_condition-method}
\title{Get prefix.}
\usage{
getPrefix(object, ...)

\S4method{getPrefix}{compartment_bioavailability}(object, ...)

\S4method{getPrefix}{compartment_lag_time}(object, ...)

\S4method{getPrefix}{compartment_infusion_duration}(object, ...)

\S4method{getPrefix}{compartment_infusion_rate}(object, ...)

\S4method{getPrefix}{compartment_initial_condition}(object, ...)
}
\arguments{
\item{object}{generic object}

\item{...}{e.g. dest='mrgsolve'}
}
\value{
the prefix of this object
}
\description{
Get prefix.
}
