% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{getUncertainty}
\alias{getUncertainty}
\title{Get uncertainty on the parameters.}
\usage{
getUncertainty(object, name)
}
\arguments{
\item{object}{generic object}
}
\value{
data frame with standard error (se) and relative standard error (rse%) columns
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`getUncertainty()` is deprecated in favor of `get_uncertainty()`.
}
\keyword{internal}
