% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/parameter.R, R/parameters.R,
%   R/campsis_model.R
\name{getUncertainty}
\alias{getUncertainty}
\alias{getUncertainty,parameter-method}
\alias{getUncertainty,parameters-method}
\alias{getUncertainty,campsis_model-method}
\title{Get uncertainty on the parameters.}
\usage{
getUncertainty(object, ...)

\S4method{getUncertainty}{parameter}(object, varcov, ...)

\S4method{getUncertainty}{parameters}(object, ...)

\S4method{getUncertainty}{campsis_model}(object, ...)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments}

\item{varcov}{variance covariance matrix}
}
\value{
data frame with standard error (se) and relative standard error (rse%) columns
}
\description{
Get uncertainty on the parameters.
}
