% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{hasEquation}
\alias{hasEquation}
\title{Has equation.}
\usage{
hasEquation(object, lhs)
}
\arguments{
\item{object}{object containing code record(s)}

\item{lhs}{left-hand-side variable to search}
}
\value{
TRUE or FALSE
}
\description{
Has equation.
}
