% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{indexOf}
\alias{indexOf}
\title{Get the index of an element in list.}
\usage{
indexOf(object, x)
}
\arguments{
\item{object}{list object}

\item{x}{element to know the index}
}
\value{
index of this element
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`indexOf()` is deprecated in favor of `index_of()`.
}
\keyword{internal}
