% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment_infusion_duration.R
\name{InfusionDuration}
\alias{InfusionDuration}
\title{Create an infusion duration.}
\usage{
InfusionDuration(compartment, rhs = "")
}
\arguments{
\item{compartment}{compartment index}

\item{rhs}{right-hand side part of the equation}
}
\value{
details about infusion duration
}
\description{
Create an infusion duration.
}
