% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{is_matrix_positive_definite}
\alias{is_matrix_positive_definite}
\title{Is matrix positive definite}
\usage{
is_matrix_positive_definite(matrix)
}
\arguments{
\item{matrix}{matrix to check}
}
\description{
Is matrix positive definite
}
