% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_strict_record_delimiter}
\alias{is_strict_record_delimiter}
\title{Is strict record delimiter. A strict record delimiter is any line starting
with [...] and followed by nothing but spaces or a possible comment.}
\usage{
is_strict_record_delimiter(line)
}
\arguments{
\item{line}{any line, single character value}
}
\value{
a logical value
}
\description{
Is strict record delimiter. A strict record delimiter is any line starting
with [...] and followed by nothing but spaces or a possible comment.
}
