% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R, R/code_record.R
\name{length,pmx_list-method}
\alias{length,pmx_list-method}
\alias{length,statement_record-method}
\title{Return the length of this list.}
\usage{
\S4method{length}{pmx_list}(x)

\S4method{length}{statement_record}(x)
}
\arguments{
\item{x}{list object}
}
\description{
Return the length of this list.
}
