% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/parameter.R, R/campsis_model.R
\name{loadFromJSON}
\alias{loadFromJSON}
\alias{loadFromJSON,theta,json_element-method}
\alias{loadFromJSON,omega,json_element-method}
\alias{loadFromJSON,sigma,json_element-method}
\alias{loadFromJSON,campsis_model,json_element-method}
\alias{loadFromJSON,campsis_model,character-method}
\title{Fill-in S4 object from the JSON content.}
\usage{
loadFromJSON(object, json)

\S4method{loadFromJSON}{theta,json_element}(object, json)

\S4method{loadFromJSON}{omega,json_element}(object, json)

\S4method{loadFromJSON}{sigma,json_element}(object, json)

\S4method{loadFromJSON}{campsis_model,json_element}(object, json)

\S4method{loadFromJSON}{campsis_model,character}(object, json)
}
\arguments{
\item{object}{pre-initiated S4 object}

\item{json}{JSON (usually a list)}
}
\value{
the loaded S4 object
}
\description{
Fill-in S4 object from the JSON content.
}
