% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{mapS4SlotsToJSONProperties}
\alias{mapS4SlotsToJSONProperties}
\title{Map S4 slots to JSON properties.}
\usage{
mapS4SlotsToJSONProperties(
  object,
  add_type = TRUE,
  optional = NULL,
  ignore = NULL
)
}
\value{
Campsis dataset
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`mapS4SlotsToJSONProperties()` is deprecated in favor of `map_s4_slots_to_json_properties()`.
}
