% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_interface.R
\name{mapS4SlotsToJSONProperties}
\alias{mapS4SlotsToJSONProperties}
\title{Map S4 slots to JSON properties.}
\usage{
mapS4SlotsToJSONProperties(
  object,
  add_type = TRUE,
  optional = NULL,
  ignore = NULL
)
}
\arguments{
\item{object}{S4 object}

\item{add_type}{add type as a property, TRUE by default}

\item{ignore}{slots to be ignored}

\item{properties}{properties that are optional in JSON, character vector}
}
\value{
a JSON object ready to be serialised
}
\description{
Map S4 slots to JSON properties.
}
