% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{maxIndex}
\alias{maxIndex}
\title{Max index.}
\usage{
maxIndex(object, type)
}
\arguments{
\item{object}{generic object}

\item{type}{parameter type: theta, omega or sigma}
}
\value{
filtered object
}
\description{
Max index.
}
