% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{maxIndex}
\alias{maxIndex}
\alias{maxIndex,parameters-method}
\title{Max index.}
\usage{
maxIndex(object)

\S4method{maxIndex}{parameters}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
max index
}
\description{
Max index.
}
