% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{minIndex}
\alias{minIndex}
\title{Min index.}
\usage{
minIndex(object)
}
\arguments{
\item{object}{generic object}
}
\value{
min index
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`minIndex()` is deprecated in favor of `min_index()`.
}
\keyword{internal}
