% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{minIndex}
\alias{minIndex}
\alias{minIndex,parameters-method}
\title{Min index.}
\usage{
minIndex(object)

\S4method{minIndex}{parameters}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
min index
}
\description{
Min index.
}
