% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{mrgsolveBlock}
\alias{mrgsolveBlock}
\title{Convert code record for mrgsolve.}
\usage{
mrgsolveBlock(record, init = NULL)
}
\arguments{
\item{record}{code record}

\item{init}{name of mrgsolve block}
}
\value{
translated record for mrgsolve
}
\description{
Convert code record for mrgsolve.
}
