% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{mrgsolveParam}
\alias{mrgsolveParam}
\title{Get the parameters block for mrgsolve.}
\usage{
mrgsolveParam(model)
}
\arguments{
\item{model}{PMX model}
}
\value{
character vector, each value is a line or character(0) if no param
}
\description{
Get the parameters block for mrgsolve.
}
