% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve_conversion.R
\name{mrgsolveTable}
\alias{mrgsolveTable}
\title{Get the TABLE block for mrgsolve.}
\usage{
mrgsolveTable(model)
}
\arguments{
\item{model}{CAMPSIS model}
}
\value{
TABLE block if at least one line in error record, character(0) otherwise
}
\description{
Get the TABLE block for mrgsolve.
}
