% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{Omega}
\alias{Omega}
\title{Create an OMEGA parameter.}
\usage{
Omega(
  name = NA,
  index = NA,
  index2 = NA,
  value = NA,
  fix = FALSE,
  type = NULL,
  same = NA
)
}
\arguments{
\item{name}{parameter name, e.g. CL (prefix OMEGA will be added automatically)}

\item{index}{parameter index}

\item{index2}{second parameter index}

\item{value}{parameter value}

\item{fix}{parameter was fixed in estimation, logical value}

\item{type}{variance type: 'var', 'sd', 'covar', 'cv' or 'cv\%'}

\item{same}{NA by default, FALSE for first OMEGA followed by 'SAME' OMEGA's, TRUE for to 'SAME' OMEGA's}
}
\value{
an OMEGA parameter
}
\description{
Create an OMEGA parameter.
}
