% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parser.R
\name{parseIfStatement}
\alias{parseIfStatement}
\title{Parse IF-statement.
Assumption: isIfStatement method already called and returned TRUE.}
\usage{
parseIfStatement(line, comment = as.character(NA))
}
\arguments{
\item{line}{IF-statement as single character string value, comment omitted}

\item{comment}{any comment, NA by default}
}
\value{
an IF statement object
}
\description{
Parse IF-statement.
Assumption: isIfStatement method already called and returned TRUE.
}
