% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_element_position.R
\name{Position}
\alias{Position}
\title{Element position in list.}
\usage{
Position(x, after = TRUE)
}
\arguments{
\item{x}{either an integer position (useful to add an element in a code record at a specified position) or
an model element (element can be a model statement or a code record)}

\item{after}{element to be added will be added after x (if after is TRUE) or before x (if after is FALSE)}
}
\value{
a position object
}
\description{
Element position in list.
}
