% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campsis_model.R
\name{read.pmxmod}
\alias{read.pmxmod}
\title{Read a Campsis model (deprecated).}
\usage{
read.pmxmod(file)
}
\arguments{
\item{file}{path to folder}
}
\value{
a Campsis model
}
\description{
Read a Campsis model (deprecated).
}
\keyword{internal}
