% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_records.R
\name{removeEquation}
\alias{removeEquation}
\title{Remove equation from code records.}
\usage{
removeEquation(object, lhs)
}
\arguments{
\item{object}{object containing code records}

\item{lhs}{left hand side variable}
}
\value{
updated code records
}
\description{
Remove equation from code records.
}
