% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/code_record.R, R/code_records.R,
%   R/pmx_model.R
\name{removeEquation}
\alias{removeEquation}
\alias{removeEquation,code_record,character-method}
\alias{removeEquation,code_records,character-method}
\alias{removeEquation,pmx_model,character-method}
\title{Remove equation from code record.}
\usage{
removeEquation(object, lhs)

\S4method{removeEquation}{code_record,character}(object, lhs)

\S4method{removeEquation}{code_records,character}(object, lhs)

\S4method{removeEquation}{pmx_model,character}(object, lhs)
}
\arguments{
\item{object}{object containing code record(s)}

\item{lhs}{left-hand-side variable}
}
\value{
updated code record
}
\description{
Remove equation from code record.
}
