% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/pattern.R, R/model_statement.R,
%   R/model_unknown_statement.R, R/model_equation.R, R/model_if_statement.R,
%   R/compartment_property.R, R/code_record.R, R/code_records.R,
%   R/campsis_model.R
\name{replaceAll}
\alias{replaceAll}
\alias{replaceAll,character,variable_pattern,character-method}
\alias{replaceAll,character,pattern,character-method}
\alias{replaceAll,model_statement,pattern,character-method}
\alias{replaceAll,unknown_statement,pattern,character-method}
\alias{replaceAll,equation,pattern,character-method}
\alias{replaceAll,if_statement,pattern,character-method}
\alias{replaceAll,compartment_property,pattern,character-method}
\alias{replaceAll,code_record,pattern,character-method}
\alias{replaceAll,code_records,pattern,character-method}
\alias{replaceAll,campsis_model,pattern,character-method}
\alias{replaceAll,campsis_model,character,character-method}
\title{Replace all occurrences in object.}
\usage{
replaceAll(object, pattern, replacement, ...)

\S4method{replaceAll}{character,variable_pattern,character}(object, pattern, replacement, ...)

\S4method{replaceAll}{character,pattern,character}(object, pattern, replacement, ...)

\S4method{replaceAll}{model_statement,pattern,character}(object, pattern, replacement, ...)

\S4method{replaceAll}{unknown_statement,pattern,character}(object, pattern, replacement, ...)

\S4method{replaceAll}{equation,pattern,character}(object, pattern, replacement, ...)

\S4method{replaceAll}{if_statement,pattern,character}(object, pattern, replacement, ...)

\S4method{replaceAll}{compartment_property,pattern,character}(object, pattern, replacement, ...)

\S4method{replaceAll}{code_record,pattern,character}(object, pattern, replacement, ...)

\S4method{replaceAll}{code_records,pattern,character}(object, pattern, replacement, ...)

\S4method{replaceAll}{campsis_model,pattern,character}(object, pattern, replacement, ...)

\S4method{replaceAll}{campsis_model,character,character}(object, pattern, replacement, ...)
}
\arguments{
\item{object}{generic object (e.g. model, code_record(s), etc.)}

\item{pattern}{pattern to be replaced}

\item{replacement}{replacement string}

\item{...}{extra arguments}
}
\value{
the same object with all occurrences replaced
}
\description{
Replace all occurrences in object.
}
