% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/code_record.R, R/code_records.R,
%   R/pmx_model.R
\name{replaceEquation}
\alias{replaceEquation}
\alias{replaceEquation,code_record,character,character-method}
\alias{replaceEquation,code_records,character,character-method}
\alias{replaceEquation,pmx_model,character,character-method}
\title{Replace equation in code records.}
\usage{
replaceEquation(object, lhs, rhs)

\S4method{replaceEquation}{code_record,character,character}(object, lhs, rhs)

\S4method{replaceEquation}{code_records,character,character}(object, lhs, rhs)

\S4method{replaceEquation}{pmx_model,character,character}(object, lhs, rhs)
}
\arguments{
\item{object}{object containing code records}

\item{lhs}{left-hand-side variable}

\item{rhs}{right-hand-side expression}
}
\value{
updated code records
}
\description{
Replace equation in code records.
}
