% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/replicated_campsis_model.R
\name{replicate}
\alias{replicate}
\alias{replicate,campsis_model,integer,replication_settings-method}
\title{Replicate generic object.}
\usage{
replicate(object, n, settings = NULL, ...)

\S4method{replicate}{campsis_model,integer,replication_settings}(object, n, settings)
}
\arguments{
\item{object}{generic object}

\item{n}{number of replicates required}

\item{settings}{settings for replication}

\item{...}{extra arguments}
}
\value{
object replicated n times
}
\description{
Replicate generic object.
}
