% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode_conversion.R
\name{rxodeCode}
\alias{rxodeCode}
\title{Get code for RxODE.}
\usage{
rxodeCode(model)
}
\arguments{
\item{model}{CAMPSIS model}
}
\value{
corresponding model code for RxODE
}
\description{
Get code for RxODE.
}
