% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{rxodeMatrix}
\alias{rxodeMatrix}
\title{Get the OMEGA/SIGMA matrix for rxode2.}
\usage{
rxodeMatrix(model, type)
}
\arguments{
\item{model}{Campsis model or Campsis parameters}

\item{type}{either omega or sigma}
}
\value{
omega/sigma named matrix
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`rxodeMatrix()` is deprecated in favor of `rxode_matrix()`.
}
\keyword{internal}
