% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode_conversion.R
\name{rxodeMatrix}
\alias{rxodeMatrix}
\title{Get the matrix for RxODE.}
\usage{
rxodeMatrix(model, type = "omega")
}
\arguments{
\item{model}{PMX model}

\item{type}{either omega or sigma}
}
\value{
named matrix
}
\description{
Get the matrix for RxODE.
}
