% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{sampleFromMultivariateNormalDistribution}
\alias{sampleFromMultivariateNormalDistribution}
\title{Sample from a multivariate normal distribution.}
\usage{
sampleFromMultivariateNormalDistribution(parameters, varcov, n, quiet)
}
\arguments{
\item{parameters}{Campsis parameters present in the variance-covariance matrix}

\item{varcov}{variance-covariance matrix}

\item{n}{number of rows to sample}

\item{quiet}{suppress messages}
}
\value{
a data frame with the sampled parameters
}
\description{
Sample from a multivariate normal distribution.
}
