% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty_utils.R
\name{sampleFromMultivariateNormalDistributionCore}
\alias{sampleFromMultivariateNormalDistributionCore}
\title{Sample more parameters.}
\usage{
sampleFromMultivariateNormalDistributionCore(n, mean, varcov)
}
\arguments{
\item{n}{numbers of rows to sample}

\item{mean}{mean values to give to the multivariate normal distribution}

\item{varcov}{variance-covariance matrix}
}
\description{
Sample more parameters.
}
