% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicated_campsis_model.R
\name{sampleOmegasSigmas}
\alias{sampleOmegasSigmas}
\title{Sample the OMEGAs and SIGMAs from inverse chi-square or inverse wishart distributions.}
\usage{
sampleOmegasSigmas(model, type = "omega", n, nsub, nobs)
}
\arguments{
\item{model}{Campsis model}

\item{type}{type of parameter to sample (omega or sigma)}

\item{n}{number of rows to sample}

\item{nsub}{number of subjects in the dataset on which the model was estimated}

\item{nobs}{number of observations in the dataset on which the model was estimated}
}
\value{
a data frame with the sampled parameters
}
\description{
Sample the OMEGAs and SIGMAs from inverse chi-square or inverse wishart distributions.
}
