% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{Sigma}
\alias{Sigma}
\title{Create a SIGMA parameter.}
\usage{
Sigma(
  name = NA,
  index = NA,
  index2 = NA,
  value = NA,
  min = NA,
  max = NA,
  fix = FALSE,
  type = NULL,
  label = NA,
  comment = NA
)
}
\arguments{
\item{name}{parameter name, e.g. CL (prefix SIGMA will be added automatically)}

\item{index}{parameter index}

\item{index2}{second parameter index}

\item{value}{parameter value}

\item{min}{minimum value for this parameter when parameter uncertainty is enabled}

\item{max}{maximum value for this parameter when parameter uncertainty is enabled}

\item{fix}{parameter was fixed in estimation, logical value}

\item{type}{variance type: 'var', 'sd', 'covar', 'cv' or 'cv\%'}

\item{label}{parameter label, optional}

\item{comment}{any comment, optional}
}
\value{
a SIGMA parameter
}
\description{
Create a SIGMA parameter.
}
