% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{standardise}
\alias{standardise}
\title{Standardise.}
\usage{
standardise(object, ...)
}
\arguments{
\item{object}{generic object}

\item{...}{arguments needed for standardisation}
}
\value{
standardised object
}
\description{
Standardise.
}
