% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/parameter.R, R/parameters.R
\name{standardise}
\alias{standardise}
\alias{standardise,theta-method}
\alias{standardise,double_array_parameter-method}
\alias{standardise,parameters-method}
\title{Standardise.}
\usage{
standardise(object, ...)

\S4method{standardise}{theta}(object, ...)

\S4method{standardise}{double_array_parameter}(object, parameters = NULL, ...)

\S4method{standardise}{parameters}(object, ...)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments needed for standardisation}

\item{parameters}{the list of parameters, to be provided only if parameter type is 'cor'}
}
\value{
standardised object
}
\description{
Standardise.
}
