% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{Theta}
\alias{Theta}
\title{Create a THETA parameter.}
\usage{
Theta(name = NA, index, value = NA, fix = FALSE)
}
\arguments{
\item{name}{parameter name, e.g. CL (prefix THETA will be added automatically)}

\item{index}{parameter index}

\item{value}{parameter value}

\item{fix}{parameter was fixed in estimation, logical value}
}
\value{
a THETA parameter
}
\description{
Create a THETA parameter.
}
