% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{toString}
\alias{toString}
\title{to_string generic method.}
\usage{
toString(object, ...)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments needed for to_string conversion}
}
\value{
character value/vector
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`toString()` is deprecated in favor of `to_string()`.
}
\keyword{internal}
