% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{toString}
\alias{toString}
\title{ToString generic method}
\usage{
toString(object, ...)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments needed for toString conversion}
}
\value{
character value/vector
}
\description{
ToString generic method
}
