% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/compartment_property.R,
%   R/compartment_initial_condition.R, R/mrgsolve_model.R
\name{toString}
\alias{toString}
\alias{toString,compartment_property-method}
\alias{toString,compartment_initial_condition-method}
\alias{toString,mrgsolve_model-method}
\title{ToString generic method}
\usage{
toString(object, ...)

\S4method{toString}{compartment_property}(object, ...)

\S4method{toString}{compartment_initial_condition}(object, ...)

\S4method{toString}{mrgsolve_model}(object, ...)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments needed for toString conversion}
}
\value{
character value/vector
}
\description{
ToString generic method
}
