% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/model_unknown_statement.R,
%   R/model_comment.R, R/model_line_break.R, R/model_equation.R, R/model_ode.R,
%   R/model_if_statement.R, R/compartment.R, R/compartment_property.R,
%   R/compartment_initial_condition.R, R/mrgsolve_model.R
\name{toString}
\alias{toString}
\alias{toString,unknown_statement-method}
\alias{toString,comment-method}
\alias{toString,line_break-method}
\alias{toString,equation-method}
\alias{toString,ode-method}
\alias{toString,if_statement-method}
\alias{toString,compartment-method}
\alias{toString,compartment_property-method}
\alias{toString,compartment_initial_condition-method}
\alias{toString,mrgsolve_model-method}
\title{ToString generic method.}
\usage{
toString(object, ...)

\S4method{toString}{unknown_statement}(object, ...)

\S4method{toString}{comment}(object, ...)

\S4method{toString}{line_break}(object, ...)

\S4method{toString}{equation}(object, ...)

\S4method{toString}{ode}(object, ...)

\S4method{toString}{if_statement}(object, ...)

\S4method{toString}{compartment}(object, ...)

\S4method{toString}{compartment_property}(object, ...)

\S4method{toString}{compartment_initial_condition}(object, ...)

\S4method{toString}{mrgsolve_model}(object, ...)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments needed for toString conversion}
}
\value{
character value/vector
}
\description{
ToString generic method.
}
