% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campsis_model.R
\name{updateCompartments}
\alias{updateCompartments}
\title{Update compartments list from the persisted records.
Exported especially for package pmxtran. However, this method should not be called.}
\usage{
updateCompartments(model)
}
\arguments{
\item{model}{CAMPSIS model}
}
\value{
an updated CAMPSIS model, with an updated compartments list
}
\description{
Update compartments list from the persisted records.
Exported especially for package pmxtran. However, this method should not be called.
}
