% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{variablePattern}
\alias{variablePattern}
\title{Return the variable pattern.}
\usage{
variablePattern()
}
\value{
pattern (regular expression)
}
\description{
Return the variable pattern.
}
\keyword{internal}
