% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/parameters.R, R/code_records.R,
%   R/campsis_model.R
\name{write}
\alias{write}
\alias{write,parameters,character-method}
\alias{write,code_records,character-method}
\alias{write,campsis_model,character-method}
\alias{write,json_element,character-method}
\title{Write generic object to files.}
\usage{
write(object, file, ...)

\S4method{write}{parameters,character}(object, file, ...)

\S4method{write}{code_records,character}(object, file, ...)

\S4method{write}{campsis_model,character}(object, file, ...)

\S4method{write}{json_element,character}(object, file, ...)
}
\arguments{
\item{object}{generic object}

\item{file}{path of the output file or directory}

\item{...}{extra arguments}
}
\value{
logical value, TRUE for success, FALSE for failure
}
\description{
Write generic object to files.
}
