% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{write}
\alias{write}
\title{Write generic object.}
\usage{
write(object, file, ...)
}
\arguments{
\item{object}{generic object}

\item{file}{path of the output dir or ZIP filename}

\item{...}{extra arguments}
}
\description{
Write generic object.
}
